refactor(cli): split cli.py into a package of subcommand family modules
Eight verb families (refute, ruling, amend, lifecycle, team, hooks,
skill, audit) move out of the 5,388-line cli.py into their own modules
under daimon_briefing/cli/, plus _ledger.py for the record printers and
channel resolver both ledger polarities share. Pure move: handlers and
register() reach shared helpers through the package namespace so the
cli.<name> monkeypatch seam keeps landing on moved code; every moved
name is re-imported in __init__; cli/__main__.py preserves the
python -m spelling; the console entry point is unchanged. The
.scars/0025 path anchor follows the code. The relations import guard
now walks the package tree (its cli.py allowlist entry went stale).

Refs #708
