+++ plugin/daimon_briefing/cli/__init__.py
@@ -31,14 +31,14 @@ import traceback
 from datetime import datetime, timedelta, timezone
 from pathlib import Path
 
-from . import amendments, anchor, briefing, capture, carry, config, configure, harvest, inspector, ledger, llm, normalize, privacy, provenance, recall, receipts, redact, refutations, relations, render, schema, serializer, store, teamsync, transcript, worldcheck
-from . import __version__
+from .. import amendments, anchor, briefing, capture, carry, config, configure, harvest, inspector, ledger, llm, normalize, privacy, provenance, recall, receipts, redact, refutations, relations, render, schema, serializer, store, teamsync, transcript, worldcheck  # noqa: F401 — several are re-exported for compat only (#708): `cli.<name>` is a stable seam
+from .. import __version__
 
 # The serialize.log ledger subsystem lives in ledger.py (#147 + #162, pure
 # moves). EVERY moved name is re-imported here — including the ones cli.py no
 # longer calls itself — because `cli.<name>` is a stable seam: tests and host
 # hooks resolve the ledger through this module.
-from .ledger import (
+from ..ledger import (
     AUTO_BRIEF_HOSTS,
     _HEAL_SKIP_REASON,  # noqa: F401 — re-exported for compat
     _HEAL_TRANSCRIPT_RE,  # noqa: F401 — re-exported for compat
@@ -880,752 +880,85 @@ def _cmd_projects(args) -> int:
 # ---- resolve/log: zero-LLM append-only event writers (#102) ----
 
 
-def _cmd_resolve(args) -> int:
-    """Append a resolution event for ONE checkpoint item (#102). Exact id
-    first; else a fuzzy query that must match uniquely — an ambiguous bind
-    is refused with candidates listed, because a wrong bind silently
-    suppresses a live memory (the false-merge lesson, #13).
-
-    `--dry-run` runs this SAME bind and stops right before the event write
-    (#304) — the matcher a preview would use is otherwise inherently
-    different from the one doing the write (recall: FTS5 over history
-    across checkpoints; resolve: carry._same_item over this checkpoint
-    only), and a preview that disagrees with the writer is worse than none.
-    Ambiguous/no-match refusals return before this point either way, so
-    their output is identical with or without the flag.
-
-    #480 slice 2 — the agent write path: `--by agent --evidence "<quote>"`
-    appends a `resolving-candidate` event, `source="agent"`, instead of the
-    human path's immediate `resolved`/`source="cli"`. Evidence is mandatory
-    with `--by agent` (mirrors #103 reverify's evidence gate on the reopen
-    side) — a call without it is refused before any checkpoint I/O, nothing
-    written, logged as `resolve:no-evidence` (extends #303). `--evidence`
-    without `--by agent` is rejected too: it is not a `--note` synonym on
-    the human path. The bind (exact id, unique fuzzy match, ambiguous/
-    no-match refusal, --dry-run) is identical on both paths — only what gets
-    written at the end differs. `resolving-candidate` is deliberately NOT
-    'resolved': store.is_resolved/_tie_rank exempt it exactly as they exempt
-    #14's supersede-candidate, so an agent's claim never withholds the item
-    on its own (the #13 false-merge lesson) until slice 3's serializer
-    verifies the quote or a human confirms."""
-    by_agent = getattr(args, "by", None) == "agent"
-    raw_evidence = getattr(args, "evidence", None)
-    if raw_evidence is not None and not by_agent:
-        print("--evidence requires --by agent — the human path vouches by "
-              "calling resolve directly, no evidence needed")
-        return 1
-    evidence = (raw_evidence or "").strip()
-    if by_agent and not evidence:
-        # Same #303 stance as the ambiguous/no-match refusals below: a
-        # refused attempt must leave its own trace, so "no agent ever tried"
-        # and "an agent tried and was refused for lacking evidence" stay
-        # distinguishable in `daimon stats`.
-        _note_usage("resolve:no-evidence")
-        print("--by agent requires --evidence \"<verbatim transcript quote>\" "
-              "— refused, nothing written")
-        return 1
-    project = _resolve_project(args.project)
-    checkpoint = store.read_latest(project_dir=project, fallback=False)
-    if not isinstance(checkpoint, dict):
-        print("no checkpoint for this project yet — nothing to resolve")
-        return 1
-    items = []
-    for section, key in store._ITEM_LISTS:
-        for item in ((checkpoint.get(section) or {}).get(key) or []):
-            if isinstance(item, dict) and item.get("id"):
-                items.append((key, item))
-    target = next((it for _, it in items if it["id"] == args.target), None)
-    if target is None:
-        texts = [str(it.get("text") or "") for _, it in items]
-        generic = carry._generic_terms(texts)
-        hits = [(key, it) for key, it in items
-                if carry._same_item(args.target, str(it.get("text") or ""), generic)]
-        if len(hits) == 1:
-            target = hits[0][1]
-        else:
-            # #303: a refused attempt must leave its own trace — otherwise
-            # "no agent ever tried resolve" and "an agent tried and was
-            # refused" are indistinguishable in `daimon stats`, and the two
-            # have opposite fixes (teaching vs UX).
-            _note_usage("resolve:no-match" if not hits else "resolve:ambiguous")
-            label = "no item matches" if not hits else "ambiguous — matches"
-            print(f"{label} {args.target!r}; candidates:")
-            listing = hits or items
-            for key, it in listing:
-                print(f"  {it['id']}  [{key}] {it.get('text', '')}")
-            print("resolve by exact id: daimon resolve <id>")
-            return 1
-    # #480 slice 2: the agent path writes a candidate status, credited to
-    # source="agent" — NOT args.status/"cli", which stay exactly what the
-    # human path has always written (byte-identical human behavior).
-    effective_status = "resolving-candidate" if by_agent else args.status
-    if getattr(args, "dry_run", False):
-        # A distinct tag, not "resolve": nothing was written, so folding this
-        # into the success counter would inflate it with attempts that never
-        # touched the ledger — corrupting the exact refusal-rate signal #303
-        # exists to expose. Not silent either: the tag still shows up in
-        # `daimon stats` usage counts, just apart from resolve/resolve:*.
-        _note_usage("resolve:dry-run")
-        print(f"would resolve {target['id']}: {target.get('text', '')} [{effective_status}]")
-        return 0
-    ok = store.append_event(
-        target["id"], effective_status,
-        note=(evidence if by_agent else (args.note or "")),
-        source=("agent" if by_agent else "cli"),
-        project_dir=project, item_text=str(target.get("text") or ""))
-    if not ok:
-        print("event not written (daimon disabled or project unknown)")
-        return 1
-    if by_agent:
-        _note_usage("resolve:agent")
-        print(f"claim recorded {target['id']}: {target.get('text', '')} "
-              "— pending verification at session end")
-        return 0
-    _note_usage("resolve")
-    print(f"resolved {target['id']}: {target.get('text', '')} [{args.status}]")
-    return 0
-
-
-# ---- refute: project-level negative-knowledge ledger (#573) ----
-
-
-def _refutation_json(record) -> str:
-    """JSON for one folded record or a list of them.
-
-    #576: every record carries `evidence_status: "cited"`.  `evidence` holds
-    typed source strings that were shape-checked and redacted on the way in and
-    never resolved — daimon does not open them, does not confirm the referent
-    exists, and does not judge whether it entails the verdict.  The text
-    renderer says so in a parenthetical; without this key a machine consumer
-    read sourced-looking strings under a key called `evidence` with nothing to
-    contradict the obvious reading.  The value is a constant today because
-    `cited` is the only status the ledger can currently earn (#581 would add a
-    resolved one).
-    """
-    def stamped(row: dict) -> dict:
-        return {**row, "evidence_status": "cited"}
-    payload = ([stamped(row) for row in record]
-               if isinstance(record, list) else stamped(record))
-    return json.dumps(payload, ensure_ascii=False, sort_keys=True)
-
-
-def _print_refutation(record: dict, *, detailed: bool = False,
-                      tag: bool = False) -> None:
-    state = record.get("state") or "candidate"
-    activation = record.get("activation") or f"{record.get('asserted_by', '?')}-proposed"
-    mark = "✗" if state == "active" else ("×" if state == "overturned" else "?")
-    word = "refutation " if tag else ""
-    print(f"[{word}{mark} {state} · {activation}] {record['refutation_id']}  "
-          f"{record.get('subject', '')}")
-    if not detailed:
-        return
-    print(f"  Verdict: {record.get('verdict', '')}")
-    print(f"  Scope: {record.get('scope', '')}")
-    anchors = record.get("anchors") or []
-    if anchors:
-        print(f"  Anchors: {', '.join(anchors)}")
-    revisit = record.get("revisit_when") or ""
-    if revisit:
-        print(f"  Revisit when: {revisit}")
-    evidence = record.get("evidence") or []
-    for item in evidence:
-        print(f"  Evidence: {item}")
-    if evidence:
-        # #576: Evidence sits in the same Label: value register as Provenance
-        # and Authority, which ARE derived from recorded lifecycle facts.  The
-        # source string is shape-checked and never resolved, so say so here —
-        # this is the only surface a reader of `refute show` actually reads.
-        print("  (evidence sources are recorded as cited; "
-              "daimon does not verify them)")
-    print(f"  Provenance: asserted by {record.get('asserted_by', '?')} "
-          f"({record.get('asserted_author') or 'unknown'})")
-    if record.get("activation"):
-        print(f"  Authority: {record['activation']} "
-              f"({record.get('activation_author') or 'unknown'})")
-    pending = record.get("overturn_proposed")
-    if isinstance(pending, dict):
-        print(f"  Overturn proposed by {pending.get('by', '?')} — still active")
-
-
-def _refute_channel(args) -> str:
-    """The channel this invocation actually arrived through.
-
-    `--by agent` is a self-declaration of the NARROWER authority, mirroring
-    `resolve`, where the human path is likewise the ABSENCE of the flag. A
-    human path has to show an interactive terminal, and the CLI can mint
-    nothing stronger: `ui` and `signed` are in-process-only, because a channel
-    an agent can reach by shelling out is the deleted `--by human` renamed.
-    """
-    if getattr(args, "by", None) == "agent":
-        return "cli-agent"
-    if not sys.stdin.isatty():
-        raise refutations.RefutationError(
-            "this is the human path and there is no interactive terminal; "
-            "pass --by agent to record a candidate, or run it from a terminal")
-    return "cli-tty"
-
-
-def _print_ruling(record: dict, *, detailed: bool = False,
-                  tag: bool = False) -> None:
-    """#693: a ruling renders its VERDICT (the rule text) and never the
-    refutation's ✗ glyph; an overturned ruling reads "retired" (label only,
-    the state vocabulary is unchanged); and text authored by a non-human
-    channel is labeled as such even after human ratification."""
-    state = record.get("state") or "candidate"
-    shown_state = "retired" if state == "overturned" else state
-    activation = (record.get("activation")
-                  or f"{record.get('asserted_by', '?')}-proposed")
-    authored = record.get("text_authored_by")
-    if state == "active" and authored and authored != "human":
-        activation = f"{authored}-written, {activation}"
-    mark = "§" if state == "active" else ("×" if state == "overturned" else "?")
-    word = "ruling " if tag else ""
-    print(f"[{word}{mark} {shown_state} · {activation}] "
-          f"{record['refutation_id']}  {record.get('verdict', '')}")
-    if not detailed:
-        return
-    print(f"  Governs: {record.get('subject', '')}")
-    print(f"  Scope: {record.get('scope', '')}")
-    anchors = record.get("anchors") or []
-    if anchors:
-        print(f"  Anchors: {', '.join(anchors)}")
-    revisit = record.get("revisit_when") or ""
-    if revisit:
-        print(f"  Revisit when: {revisit}")
-    for item in record.get("evidence") or []:
-        print(f"  Evidence: {item}")
-    proposal = record.get("revision_proposed")
-    if proposal:
-        print(f"  Pending revision proposal ({proposal.get('by', '?')}): "
-              f"{proposal.get('verdict') or proposal.get('subject') or ''}")
-    retirement = record.get("overturn_proposed")
-    if retirement:
-        print(f"  Pending retirement proposal ({retirement.get('by', '?')})")
-
-
-def _refuse_ruling_id(record, verb: str) -> bool:
-    """#693: one conversation per record — `refute` verbs refuse ruling ids
-    with a pointer, and vice versa."""
-    if record is not None and record.get("polarity") == "ruling":
-        print(f"{record['refutation_id']} is a ruling; use "
-              f"`daimon ruling {verb}`")
-        return True
-    return False
-
-
-def _cmd_ruling_propose(args) -> int:
-    project = _resolve_project(args.project)
-    try:
-        ruling_id = refutations.assert_ruling(
-            subject=args.subject, verdict=args.verdict, scope=args.scope,
-            evidence=args.evidence, channel=_refute_channel(args),
-            anchors=args.anchor,
-            revisit_when=args.revisit_when or "", ratified=args.ratify,
-            project_dir=project)
-    except refutations.RefutationError as exc:
-        print(f"ruling not recorded: {exc}")
-        return 1
-    record = refutations.get(ruling_id, project_dir=project)
-    _note_usage("ruling:propose")
-    if args.json:
-        print(_refutation_json(record))
-    else:
-        _print_ruling(record, detailed=True)
-        if record["state"] == "candidate":
-            if getattr(args, "by", None) != "agent":
-                print(f"  Next: daimon ruling ratify {ruling_id} "
-                      f"--project {project}")
-            else:
-                print("  Candidate recorded. Activation requires an explicit "
-                      "human decision.")
-    return 0
-
-
-def _cmd_ruling_ratify(args) -> int:
-    project = _resolve_project(args.project)
-    try:
-        channel = _refute_channel(args)
-    except refutations.RefutationError as exc:
-        print(f"ruling not ratified: {exc}")
-        return 1
-    record = refutations.get(args.ruling_id, project_dir=project)
-    if record is None:
-        print(f"unknown ruling: {args.ruling_id}")
-        return 1
-    if record.get("polarity") != "ruling":
-        print(f"{args.ruling_id} is a refutation; use `daimon refute ratify`")
-        return 1
-    if channel != "cli-tty":
-        # Never invoke a writer to harvest its error string.
-        print("ruling not ratified: ratification requires a human channel; "
-              f"this call arrived through {channel!r}")
-        return 1
-    # Ratification is a signature, not an id-typing exercise: print the FULL
-    # text, disclose the render consequence, and bind the append to the key
-    # of the text DISPLAYED — the fold refuses to activate any other text.
-    # Under --json the ceremony goes to stderr so stdout stays parseable.
-    ceremony = sys.stderr if args.json else sys.stdout
-    print("About to ratify this ruling:", file=ceremony)
-    if not args.json:
-        _print_ruling(record, detailed=True)
-    else:
-        print(f"  {record.get('verdict', '')}", file=ceremony)
-    print("  This text will render into every future session for this "
-          "project.", file=ceremony)
-    displayed_key = normalize.content_key(record.get("verdict") or "")
-    answer = input("Ratify? [y/N]: ").strip().casefold()
-    if answer not in ("y", "yes"):
-        print("not ratified")
-        return 1
-    try:
-        refutations.ratify(args.ruling_id, channel=channel,
-                           note=args.note or "", verdict_key=displayed_key,
-                           project_dir=project)
-    except refutations.RefutationError as exc:
-        print(f"ruling not ratified: {exc}")
-        return 1
-    record = refutations.get(args.ruling_id, project_dir=project)
-    _note_usage("ruling:ratify")
-    if record["state"] != "active":
-        print("not activated: the text changed during confirmation; "
-              "re-run to review the current text")
-        return 1
-    if args.json:
-        print(_refutation_json(record))
-    else:
-        _print_ruling(record, detailed=True)
-    return 0
-
-
-def _cmd_ruling_revise(args) -> int:
-    project = _resolve_project(args.project)
-    record = refutations.get(args.ruling_id, project_dir=project)
-    if record is None:
-        print(f"unknown ruling: {args.ruling_id}")
-        return 1
-    if record.get("polarity") != "ruling":
-        print(f"{args.ruling_id} is a refutation; use `daimon refute revise`")
-        return 1
-    if args.ratify:
-        # The ratification ceremony (full text, disclosure, confirm, key
-        # binding) lives in ONE place; a revise flag walking around it would
-        # activate text the human was never shown.
-        print("ruling revise does not activate; revise the candidate, then "
-              f"run `daimon ruling ratify {args.ruling_id}`")
-        return 1
-    try:
-        channel = _refute_channel(args)
-    except refutations.RefutationError as exc:
-        print(f"ruling not revised: {exc}")
-        return 1
-    if (record["state"] == "active" and channel == "cli-tty"
-            and (args.verdict is not None or args.subject is not None)):
-        # Rewriting what renders is the same power ratification has, and it
-        # earns trust the same way: show the change, disclose, confirm.
-        print("About to change the ACTIVE text of this ruling:")
-        _print_ruling(record, detailed=True)
-        if args.verdict is not None:
-            print(f"  New text: {args.verdict}")
-        if args.subject is not None:
-            print(f"  New governs: {args.subject}")
-        print("  This text will render into every future session for this "
-              "project.")
-        answer = input("Apply? [y/N]: ").strip().casefold()
-        if answer not in ("y", "yes"):
-            print("not revised")
-            return 1
-    anchors = args.anchor if args.anchor is not None else None
-    try:
-        refutations.revise(
-            args.ruling_id, channel=channel,
-            evidence=args.evidence, subject=args.subject,
-            verdict=args.verdict, scope=args.scope,
-            anchors=anchors, revisit_when=args.revisit_when,
-            ratified=False, project_dir=project)
-    except refutations.RefutationError as exc:
-        print(f"ruling not revised: {exc}")
-        return 1
-    record = refutations.get(args.ruling_id, project_dir=project)
-    _note_usage("ruling:revise")
-    if args.json:
-        print(_refutation_json(record))
-    else:
-        _print_ruling(record, detailed=True)
-        if record.get("revision_proposed"):
-            print("  Proposal recorded; the active ruling is untouched "
-                  "until a human verdict.")
-        elif record["state"] == "candidate":
-            print("  Revision is not load-bearing until explicit human "
-                  "ratification.")
-    return 0
-
-
-def _cmd_ruling_retire(args) -> int:
-    project = _resolve_project(args.project)
-    try:
-        event = refutations.retire(
-            args.ruling_id, channel=_refute_channel(args),
-            evidence=args.evidence or (),
-            note=args.note or "", project_dir=project)
-    except refutations.RefutationError as exc:
-        print(f"ruling not retired: {exc}")
-        return 1
-    record = refutations.get(args.ruling_id, project_dir=project)
-    _note_usage("ruling:retire")
-    if args.json:
-        print(_refutation_json(record))
-    else:
-        _print_ruling(record, detailed=True)
-        if event == "overturn-proposed":
-            print("  Retirement proposed; the ruling stands until a human "
-                  "verdict.")
-    return 0
-
-
-def _cmd_ruling_list(args) -> int:
-    project = _resolve_project(args.project)
-    rows = refutations.listing(states=set(args.state or refutations.STATES),
-                               polarity="ruling", project_dir=project)
-    _note_usage("ruling:list")
-    if args.json:
-        print(_refutation_json(rows))
-        active_j = sum(1 for r in rows if r.get("state") == "active")
-        cap_j = config.ruling_cap()
-        if active_j > cap_j:
-            # stderr so the JSON stays parseable and machines still learn it.
-            print(f"over cap: {active_j} active vs cap {cap_j}",
-                  file=sys.stderr)
-        return 0
-    if not rows:
-        print("no rulings for this project")
-        return 0
-    for row in rows:
-        _print_ruling(row)
-    # The cap binds ACTIVATION; a lowered DAIMON_RULING_CAP leaves the
-    # excess active, so the over-cap state must be visible somewhere.
-    active = sum(1 for r in rows if r.get("state") == "active")
-    cap = config.ruling_cap()
-    if active > cap:
-        print(f"over cap: {active} active vs cap {cap} — retire one, or "
-              "raise DAIMON_RULING_CAP deliberately")
-    return 0
-
-
-def _cmd_ruling_show(args) -> int:
-    project = _resolve_project(args.project)
-    record = refutations.get(args.ruling_id, project_dir=project)
-    if record is None:
-        print(f"unknown ruling: {args.ruling_id}")
-        return 1
-    if record.get("polarity") != "ruling":
-        print(f"{args.ruling_id} is a refutation; use `daimon refute show`")
-        return 1
-    _note_usage("ruling:show")
-    if args.json:
-        print(_refutation_json(record))
-    else:
-        _print_ruling(record, detailed=True)
-    return 0
-
-
-def _cmd_refute_add(args) -> int:
-    project = _resolve_project(args.project)
-    try:
-        ref_id = refutations.assert_refutation(
-            subject=args.subject, verdict=args.verdict, scope=args.scope,
-            evidence=args.evidence, channel=_refute_channel(args),
-            anchors=args.anchor,
-            revisit_when=args.revisit_when or "", ratified=args.ratify,
-            project_dir=project)
-    except refutations.RefutationError as exc:
-        print(f"refutation not recorded: {exc}")
-        return 1
-    record = refutations.get(ref_id, project_dir=project)
-    _note_usage("refute:add")
-    if args.json:
-        print(_refutation_json(record))
-    else:
-        _print_refutation(record, detailed=True)
-        if record["state"] == "candidate":
-            # An agent-authored candidate never gets handed its own escalation
-            # command; activation is a decision the human has to reach.
-            if getattr(args, "by", None) != "agent":
-                print(f"  Next: daimon refute ratify {ref_id} "
-                      f"--project {project}")
-            else:
-                print("  Candidate recorded. Activation requires an explicit "
-                      "human decision.")
-    return 0
-
-
-def _cmd_refute_ratify(args) -> int:
-    project = _resolve_project(args.project)
-    if _refuse_ruling_id(refutations.get(args.refutation_id,
-                                         project_dir=project), "ratify"):
-        return 1
-    try:
-        refutations.ratify(args.refutation_id, channel=_refute_channel(args),
-                           note=args.note or "", project_dir=project)
-    except refutations.RefutationError as exc:
-        print(f"refutation not ratified: {exc}")
-        return 1
-    record = refutations.get(args.refutation_id, project_dir=project)
-    _note_usage("refute:ratify")
-    if args.json:
-        print(_refutation_json(record))
-    else:
-        _print_refutation(record, detailed=True)
-    return 0
-
-
-def _cmd_refute_revise(args) -> int:
-    project = _resolve_project(args.project)
-    if _refuse_ruling_id(refutations.get(args.refutation_id,
-                                         project_dir=project), "revise"):
-        return 1
-    anchors = args.anchor if args.anchor is not None else None
-    try:
-        refutations.revise(
-            args.refutation_id, channel=_refute_channel(args),
-            evidence=args.evidence, subject=args.subject, verdict=args.verdict, scope=args.scope,
-            anchors=anchors, revisit_when=args.revisit_when,
-            ratified=args.ratify, project_dir=project)
-    except refutations.RefutationError as exc:
-        print(f"refutation not revised: {exc}")
-        return 1
-    record = refutations.get(args.refutation_id, project_dir=project)
-    _note_usage("refute:revise")
-    if args.json:
-        print(_refutation_json(record))
-    else:
-        _print_refutation(record, detailed=True)
-        if record["state"] == "candidate":
-            print("  Revision is not load-bearing until explicit human ratification.")
-    return 0
-
-
-def _cmd_refute_overturn(args) -> int:
-    project = _resolve_project(args.project)
-    if _refuse_ruling_id(refutations.get(args.refutation_id,
-                                         project_dir=project), "retire"):
-        return 1
-    try:
-        event = refutations.overturn(
-            args.refutation_id, channel=_refute_channel(args),
-            evidence=args.evidence,
-            note=args.note or "", project_dir=project)
-    except refutations.RefutationError as exc:
-        print(f"refutation not overturned: {exc}")
-        return 1
-    record = refutations.get(args.refutation_id, project_dir=project)
-    _note_usage("refute:overturn")
-    if args.json:
-        print(_refutation_json(record))
-    else:
-        _print_refutation(record, detailed=True)
-        if event == "overturn-proposed":
-            print("  Agent evidence recorded; the active guard remains until human ratification.")
-    return 0
-
-
-def _cmd_refute_show(args) -> int:
-    project = _resolve_project(args.project)
-    record = refutations.get(args.refutation_id, project_dir=project)
-    if record is None:
-        print(f"unknown refutation: {args.refutation_id}")
-        return 1
-    if _refuse_ruling_id(record, "show"):
-        return 1
-    _note_usage("refute:show")
-    if args.json:
-        print(_refutation_json(record))
-    else:
-        _print_refutation(record, detailed=True)
-    return 0
-
-
-def _cmd_refute_list(args) -> int:
-    project = _resolve_project(args.project)
-    rows = refutations.listing(states=set(args.state or refutations.STATES),
-                               polarity="refutation", project_dir=project)
-    _note_usage("refute:list")
-    if args.json:
-        print(_refutation_json(rows))
-    elif not rows:
-        print("no refutations for this project")
-    else:
-        for row in rows:
-            _print_refutation(row)
-    return 0
-
-
-def _cmd_refute_search(args) -> int:
-    project = _resolve_project(args.project)
-    try:
-        rows = refutations.search(
-            " ".join(args.query), project_dir=project,
-            states=set(args.state or refutations.STATES))
-    except refutations.RefutationError as exc:
-        print(f"refutation search refused: {exc}")
-        return 1
-    _note_usage("refute:search")
-    if args.json:
-        print(_refutation_json(rows))
-    elif not rows:
-        print("no matching refutations")
-    else:
-        # #693: search is the topic-addressable pull path, so it returns BOTH
-        # polarities, labelled by their own printers — filtering rulings out
-        # would leave the records that matter most with no pull surface
-        # mid-session, after the briefing has scrolled away.
-        for row in rows:
-            if row.get("polarity") == "ruling":
-                _print_ruling(row, tag=True)
-            else:
-                _print_refutation(row, tag=True)
-    return 0
-
-
-def _cmd_refute_guard(args) -> int:
-    project = _resolve_project(args.project)
-    try:
-        rows = refutations.guard(
-            " ".join(args.query), anchors=args.anchor,
-            project_dir=project)
-    except refutations.RefutationError as exc:
-        print(f"refutation guard refused: {exc}")
-        return 1
-    # #581: split by outcome and rail, the way `resolve` splits its tags. One
-    # aggregate count cannot separate a hit from a miss, so the false-veto rate
-    # the design named as its own expansion gate is uncomputable from field
-    # data. The rails also differ in precision, so the hit tag names which one
-    # fired: an aggregate cannot say which rail generates the noise.
-    _note_usage(f"refute:guard:hit:{rows[0]['guard_match']['rail']}"
-                if rows else "refute:guard:miss")
-    if args.json:
-        print(_refutation_json(rows))
-    elif not rows:
-        if not args.quiet:
-            print("no active refutation matched exact anchors or subject")
-    else:
-        # One warning is the v1 attention budget. JSON retains every exact hit
-        # for deliberation integrations that can reconcile them in batch.
-        _print_refutation(rows[0], detailed=True)
-        if len(rows) > 1:
-            print(f"  + {len(rows) - 1} more exact match(es); use --json to inspect")
-    return 0
-
-
-# ---- #691: amendment verbs — evidence-carrying state transitions ----
-
-
-def _amend_channel(args) -> str:
-    """The channel this invocation actually arrived through — the
-    `_refute_channel` contract restated for the amendment ledger (same
-    doctrine, its own error type so refusals stay per-surface)."""
-    if getattr(args, "by", None) == "agent":
-        return "cli-agent"
-    if not sys.stdin.isatty():
-        raise amendments.AmendmentError(
-            "this is the human path and there is no interactive terminal; "
-            "pass --by agent to record a candidate, or run it from a terminal")
-    return "cli-tty"
-
-
-def _cmd_amend_propose(args) -> int:
-    project = _resolve_project(args.project)
-    item_id = str(args.item_id or "").strip()
-    # Exact-id binding against the LIVE checkpoint only — an amendment
-    # describes an open item's state, so unlike forget it never reaches into
-    # prev-N surfaces, and unlike resolve it never fuzzy-matches: the id came
-    # off a rendered ` [id]` handle or it does not exist. Loop-shaped items
-    # only (#480's scope rule, restated): amending a settled decision or
-    # belief would invite exactly the state-rewriting on settled facts that
-    # BRIEFABLE_ITEM_KEYS exists to fence off, and `daimon loops` — the
-    # discovery surface this command's errors point at — lists only these.
-    checkpoint = store.read_latest(project_dir=project, fallback=False)
-    live = {
-        str(item.get("id") or "")
-        for section, key in store._ITEM_LISTS
-        if key in briefing.BRIEFABLE_ITEM_KEYS
-        for item in ((checkpoint or {}).get(section) or {}).get(key) or []
-        if isinstance(item, dict)
-    }
-    live.discard("")
-    if item_id not in live:
-        _note_usage("amend:no-match")
-        print(f"no open-loop item with id {item_id!r} — amend targets open "
-              "questions and uncertainties; `daimon loops` lists them")
-        return 1
-    if store.is_resolved(store.resolutions(project_dir=project).get(item_id)):
-        _note_usage("amend:resolved")
-        print(f"{item_id} is already resolved — an amendment describes an "
-              "OPEN item; `daimon reverify` reopens one first")
-        return 1
-    try:
-        a_id = amendments.propose(
-            item_id=item_id, change=args.change, evidence=args.evidence,
-            channel=_amend_channel(args), note=args.note or "",
-            project_dir=project)
-    except amendments.AmendmentError as exc:
-        _note_usage("amend:refused")
-        print(f"amendment not recorded: {exc}")
-        return 1
-    record = amendments.get(a_id, project_dir=project)
-    state = record["state"] if record else "candidate"
-    _note_usage("amend:agent" if getattr(args, "by", None) == "agent"
-                else "amend")
-    print(f"amendment {a_id} recorded on {item_id}: {args.change} ({state})")
-    if state == "candidate":
-        # Same posture as refute add: an agent-authored candidate is never
-        # handed its own escalation command — verification is the transcript
-        # byte-check at session end, settlement is a human's.
-        print("  evidence is byte-checked against the transcript at session "
-              "end; a human settles it earlier with `daimon amend ratify` "
-              "or `daimon amend reject`")
-    return 0
-
-
-def _cmd_amend_verdict(args) -> int:
-    project = _resolve_project(args.project)
-    verb = args.amend_cmd
-    try:
-        channel = _amend_channel(args)
-        if verb == "ratify":
-            amendments.ratify(args.amendment_id, channel=channel,
-                              project_dir=project)
-        else:
-            amendments.reject(args.amendment_id, channel=channel,
-                              note=getattr(args, "note", None) or "",
-                              project_dir=project)
-    except amendments.AmendmentError as exc:
-        _note_usage(f"amend:{verb}:refused")
-        print(f"amendment {verb} refused: {exc}")
-        return 1
-    _note_usage(f"amend:{verb}")
-    record = amendments.get(args.amendment_id, project_dir=project)
-    print(f"{args.amendment_id}: {record['state'] if record else 'unknown'}")
-    return 0
-
-
-def _cmd_amend_list(args) -> int:
-    project = _resolve_project(args.project)
-    _note_usage("amend:list")
-    rows = sorted(
-        amendments.records(project_dir=project).values(),
-        key=lambda r: (r["state"] != "candidate",
-                       r.get("updated_at") or "", r["amendment_id"]))
-    if args.json:
-        print(json.dumps(rows, ensure_ascii=False, indent=2))
-        return 0
-    if not rows:
-        print("no amendments recorded for this project")
-        return 0
-    for r in rows:
-        quote = briefing._truncate_agent_claim(r.get("evidence"))
-        print(f'{r["amendment_id"]}  {r["state"]:<9} {r["item_id"]}  '
-              f'{r["change"]}: "{quote}"')
-    return 0
+# ---- #708: the refutation/ruling/amendment verb families live in their own
+# modules. EVERY moved name is re-imported here — `cli.<name>` is a stable
+# seam: tests and host hooks resolve these through this module, and each
+# family's register() binds handlers via this namespace so a monkeypatch on
+# `cli` lands on moved code too.
+from ._ledger import (  # noqa: E402
+    _print_refutation,  # noqa: F401 — re-exported for compat
+    _print_ruling,  # noqa: F401 — re-exported for compat
+    _refutation_json,  # noqa: F401 — re-exported for compat
+    _refuse_ruling_id,  # noqa: F401 — re-exported for compat
+    _refute_channel,  # noqa: F401 — re-exported for compat
+)
+from .refute import (  # noqa: E402
+    _cmd_refute_add,  # noqa: F401 — re-exported for compat
+    _cmd_refute_guard,  # noqa: F401 — re-exported for compat
+    _cmd_refute_list,  # noqa: F401 — re-exported for compat
+    _cmd_refute_overturn,  # noqa: F401 — re-exported for compat
+    _cmd_refute_ratify,  # noqa: F401 — re-exported for compat
+    _cmd_refute_revise,  # noqa: F401 — re-exported for compat
+    _cmd_refute_search,  # noqa: F401 — re-exported for compat
+    _cmd_refute_show,  # noqa: F401 — re-exported for compat
+)
+from .ruling import (  # noqa: E402
+    _cmd_ruling_list,  # noqa: F401 — re-exported for compat
+    _cmd_ruling_propose,  # noqa: F401 — re-exported for compat
+    _cmd_ruling_ratify,  # noqa: F401 — re-exported for compat
+    _cmd_ruling_retire,  # noqa: F401 — re-exported for compat
+    _cmd_ruling_revise,  # noqa: F401 — re-exported for compat
+    _cmd_ruling_show,  # noqa: F401 — re-exported for compat
+)
+from .amend import (  # noqa: E402
+    _amend_channel,  # noqa: F401 — re-exported for compat
+    _cmd_amend_list,  # noqa: F401 — re-exported for compat
+    _cmd_amend_propose,  # noqa: F401 — re-exported for compat
+    _cmd_amend_verdict,  # noqa: F401 — re-exported for compat
+)
+from .lifecycle import (  # noqa: E402
+    _cmd_forget,  # noqa: F401 — re-exported for compat
+    _cmd_loops,  # noqa: F401 — re-exported for compat
+    _cmd_resolve,  # noqa: F401 — re-exported for compat
+    _cmd_reverify,  # noqa: F401 — re-exported for compat
+    _is_supersede_candidate,  # noqa: F401 — re-exported for compat
+)
+from .audit import (  # noqa: E402
+    _audit_item_source,  # noqa: F401 — re-exported for compat
+    _cmd_audit_privacy,  # noqa: F401 — re-exported for compat
+    _cmd_audit_quotes,  # noqa: F401 — re-exported for compat
+    _cmd_audit_quotes_deprecated,  # noqa: F401 — re-exported for compat
+    _legacy_audit_source,  # noqa: F401 — re-exported for compat
+    _load_audit_transcript,  # noqa: F401 — re-exported for compat
+    _resolve_audit_source,  # noqa: F401 — re-exported for compat
+)
+from .team import (  # noqa: E402
+    _cmd_team_init,  # noqa: F401 — re-exported for compat
+    _cmd_team_status,  # noqa: F401 — re-exported for compat
+    _cmd_team_sync,  # noqa: F401 — re-exported for compat
+)
+from .hooks import (  # noqa: E402
+    _cmd_hooks_install,  # noqa: F401 — re-exported for compat
+    _cmd_hooks_list,  # noqa: F401 — re-exported for compat
+    _cmd_hooks_status,  # noqa: F401 — re-exported for compat
+)
+from .skill import (  # noqa: E402
+    _cmd_skill_install,  # noqa: F401 — re-exported for compat
+    _cmd_skill_list,  # noqa: F401 — re-exported for compat
+    _cmd_skill_show,  # noqa: F401 — re-exported for compat
+    _cmd_skill_uninstall,  # noqa: F401 — re-exported for compat
+    _resolve_project_cwd,  # noqa: F401 — re-exported for compat
+)
+from . import (  # noqa: E402
+    amend,
+    audit,
+    hooks,
+    lifecycle,
+    refute,
+    ruling,
+    skill,
+    team,
+)
 
 
 def _relations_channel() -> str:
@@ -1695,655 +1028,6 @@ def _cmd_relations_verdict(args) -> int:
     _note_usage(f"relations:{args.verdict}")
     state = relations.records(project_dir=project)[args.relation_id]["state"]
     print(f"{args.relation_id} -> {state}")
-    return 0
-
-
-def _cmd_forget(args) -> int:
-    """Deliberate item removal (#321): append a tombstone event whose status
-    carries a content HASH, never the text — removal means the content leaves
-    the audit trail too — then rewrite the live checkpoint without the value.
-    Tombstone first (#418): the rewrite's _drop_forgotten reads the ledger, so
-    the key must land before the write scrubs sibling ids of the same value.
-    Binding is _cmd_resolve's never-guess contract verbatim: exact id
-    first, else a fuzzy query that must match exactly one item. The tombstone
-    rides the resolutions fold, so withhold, carry suppression, and the
-    recall index deletion all inherit it with no new plumbing. The rewritten
-    checkpoint re-mints its receipt, so the post-removal state is signed."""
-    project = _resolve_project(args.project)
-    checkpoint = store.read_latest(project_dir=project, fallback=False)
-    # #578: the refutation ledger is a second plaintext store, so a value can
-    # live there with no checkpoint at all. Bailing on a missing checkpoint
-    # would leave that value permanently unreachable.
-    # Every plaintext value the record has EVER carried, not only the folded
-    # ones: a revision rewrites fields, so the forgotten value can sit in an
-    # earlier row that nothing renders. #698: the field walk is the module's
-    # OWN declaration (plaintext_values, scalars only — a shared anchor or
-    # evidence token must never become a by-value target), never a hand-read
-    # `subject`. One entry per record, so a record whose old and new values
-    # both match is one hit; `text` stays the latest subject for display.
-    ledger_texts: dict[str, list[str]] = {}
-    ledger_display: dict[str, str] = {}
-    for row in refutations.events(project_dir=project):
-        ref_id = str(row.get("refutation_id") or "")
-        if not ref_id:
-            continue
-        for value in refutations.plaintext_values(row):
-            ledger_texts.setdefault(ref_id, [])
-            if value not in ledger_texts[ref_id]:
-                ledger_texts[ref_id].append(value)
-        subject = str(row.get("subject") or "")
-        if subject:
-            ledger_display[ref_id] = subject
-    # #693: the ledger holds BOTH polarities; every forget surface (pool
-    # label, dry-run, gate, receipt) names a ruling as a ruling. Snapshot
-    # taken BEFORE any deletion.
-    ledger_meta = {
-        rid: (rec.get("polarity") or "refutation", rec.get("state"))
-        for rid, rec in refutations.records(project_dir=project).items()}
-    ledger = [
-        (None, ledger_meta.get(ref_id, ("refutation", None))[0],
-         {"id": ref_id,
-          "text": ledger_display.get(ref_id, texts[0]),
-          "_texts": texts})
-        for ref_id, texts in ledger_texts.items()
-    ]
-    # #691: the amendment ledger is another plaintext store — evidence quotes
-    # and human notes. Same every-row posture as the refutation subjects
-    # above: a value can sit in any historical row of a record. The field
-    # walk is the module's OWN declaration (plaintext_values), never a
-    # hand-copied tuple, and `text` is the FIRST value (evidence before
-    # note) — the content hash downstream must never key on a note when the
-    # user named the evidence. `_target` carries the amended item's id so
-    # the fuzzy branch can drop amendment hits that merely orbit an item
-    # the query already matched.
-    amend_texts: dict[str, list[str]] = {}
-    amend_targets: dict[str, str] = {}
-    for row in amendments.events(project_dir=project):
-        # events() admits only rows whose amendment_id matched the id shape,
-        # so the key is present and non-empty by contract.
-        a_id = str(row["amendment_id"])
-        target_id = str(row.get("item_id") or "")
-        if target_id:
-            amend_targets.setdefault(a_id, target_id)
-        for value in amendments.plaintext_values(row):
-            amend_texts.setdefault(a_id, [])
-            if value not in amend_texts[a_id]:
-                amend_texts[a_id].append(value)
-    amend_pool = [
-        (None, "amendment", {"id": a_id, "text": texts[0], "_texts": texts,
-                             "_target": amend_targets.get(a_id, "")})
-        for a_id, texts in amend_texts.items()
-    ]
-    if not isinstance(checkpoint, dict) and not ledger and not amend_pool:
-        print("no checkpoint for this project yet — nothing to forget")
-        return 1
-    # Every surface this project holds, not just the live checkpoint (#419
-    # scope): a value that had been superseded still sits in prev-N and in its
-    # session file, and resolving only against `latest` made forget answer "no
-    # item matches" about plaintext that was demonstrably on disk.
-    seen_ids: set[str] = set()
-    items = []
-    for _path, section, key, item in store.items_for_project(project):
-        if item["id"] in seen_ids:
-            continue          # one item, several surfaces — not several items
-        seen_ids.add(item["id"])
-        items.append((section, key, item))
-    # Refutation ids and checkpoint `recent_decisions` ids SHARE the namespace
-    # `r-<12 hex>`, so an exact-id lookup can legitimately hit both surfaces.
-    # forget's never-guess contract decides it: an ambiguous id is refused, not
-    # resolved by preferring a store.
-    candidates = items + ledger + amend_pool
-    exact = [it for _, _, it in candidates if it["id"] == args.target]
-    target = exact[0] if len(exact) == 1 else None
-    if target is None:
-        # `_generic_terms` is a DOCUMENT-FREQUENCY statistic over texts "of one
-        # kind" (carry.py's own wording): terms shared by >= _GENERIC_DF of them
-        # are that kind's boilerplate and are subtracted from the matcher.
-        # Counting the checkpoint and the ledger together mixes two corpora, so
-        # recording refutations inflated the frequency until the query's own
-        # terms read as generic — and `forget "<text>"` stopped reaching a
-        # checkpoint item it had deleted a moment earlier, reporting "no item
-        # matches" about plaintext on disk. Each store is counted on its own
-        # and each candidate matched against its own store's vocabulary.
-        def _texts_of(it):
-            return it.get("_texts") or [str(it.get("text") or "")]
-
-        # #698 review: document frequency is per CANDIDATE, never per field.
-        # A record's subject/verdict/scope restate each other by construction,
-        # so counting them as separate documents lets one record push its own
-        # terms over the generic threshold and hide itself from its exact
-        # subject. One concatenated document per candidate keeps the statistic
-        # meaning what carry.py says it means: terms shared by >= k ITEMS.
-        pools = [(pool, carry._generic_terms(
-            [" ".join(_texts_of(it)) for _, _, it in pool]))
-            for pool in (items, ledger, amend_pool)]
-        query_key = normalize.content_key(args.target)
-
-        def _matched_value(it, generic):
-            # The value the QUERY named. The exact canonical match is a rail
-            # the GENERIC FILTER can never subtract, no matter how common the
-            # value's terms have become (#698 review) — the never-guess
-            # ambiguity gate below may still refuse it when several values
-            # match. The fuzzy fallback matches each value against the
-            # candidate's own store vocabulary.
-            for t in _texts_of(it):
-                if normalize.content_key(t) == query_key:
-                    return t
-            for t in _texts_of(it):
-                if carry._same_item(args.target, t, generic):
-                    return t
-            return None
-
-        # Hits carry the value they matched IN the tuple — a side table keyed
-        # on object identity would silently degrade to display text if a
-        # refactor ever copied a candidate dict (#698 review).
-        if len(exact) > 1:
-            hits = [(s, k, it, None) for s, k, it in candidates
-                    if it["id"] == args.target]
-        else:
-            hits = [(s, k, it, value)
-                    for pool, generic in pools
-                    for s, k, it in pool
-                    for value in [_matched_value(it, generic)]
-                    if value is not None]
-        # #691: an amendment's evidence is by construction ABOUT its item, so
-        # a query matching the item routinely fuzzy-matches the amendment too
-        # — a false-ambiguity surface, not a real second value. When an item
-        # hit exists, amendment hits that merely target one of the hit items
-        # are dropped: forgetting the item removes its amendments anyway
-        # (forget_item_id below), so nothing becomes unreachable.
-        item_hit_ids = {it["id"] for _, k, it, _ in hits
-                        if k not in ("refutation", "amendment")}
-        if item_hit_ids:
-            hits = [(s, k, it, m) for s, k, it, m in hits
-                    if not (k == "amendment"
-                            and it.get("_target") in item_hit_ids)]
-        # Ambiguity is about distinct MATCHED values, not hit count and never
-        # the display text. The same sentence carried by sibling ids, held on
-        # several surfaces, or held in both the checkpoint and the refutation
-        # ledger is one thing to forget; #418 already splices sibling ids from
-        # a single key. But two records matched on DIFFERENT values behind one
-        # shared display subject are two things — collapsing them over the
-        # display text silently under-deleted (#698 review). Only genuinely
-        # different matched values leave the user a choice, and there
-        # never-guess still refuses.
-        distinct = {normalize.content_key(
-            m if m is not None else str(it.get("text") or ""))
-            for _, _, it, m in hits}
-        if len(distinct) == 1:
-            _, _, target, matched = hits[0]
-            # #691: a record can hold several values (evidence, note,
-            # historical rows). The tombstone must key on the value the QUERY
-            # named, never on an arbitrary field — rebind `text` to the
-            # matched value before the hash below derives from it.
-            if matched is not None and matched != target.get("text"):
-                target = dict(target, text=matched)
-        else:
-            _note_usage("forget:no-match" if not hits else "forget:ambiguous")
-            label = "no item matches" if not hits else "ambiguous — matches"
-            print(f"{label} {args.target!r}; candidates:")
-            # A never-guess refusal is only useful if the user can make the
-            # choice: when the matched value differs from the display text,
-            # show it, or two candidates separated by their verdicts render
-            # as identical lines (#698 review).
-            rows = hits or [(s, k, it, None) for s, k, it in candidates]
-            for _, key, it, m in rows:
-                line = f"  {it['id']}  [{key}] {it.get('text', '')}"
-                if m is not None and m != it.get("text"):
-                    line += f" — matched: {m}"
-                print(line)
-            print("forget by exact id: daimon forget <id>")
-            return 1
-    if getattr(args, "dry_run", False):
-        _note_usage("forget:dry-run")
-        # #698 review: forget is irreversible and this preview is the only
-        # pre-deletion check. The deleters reach EVERY record whose any
-        # declared field folds to the value's key — including list fields the
-        # selector deliberately never offers — so the preview enumerates that
-        # reach non-destructively instead of naming only the selected target.
-        value_key = normalize.content_key(str(target.get("text") or ""))
-        # The checkpoint splice takes every item holding the value, whatever
-        # its id — and amendments die with their item BY ID, carrying prose
-        # that is not the forgotten value, so both must be previewed too
-        # (#698 review). Same computation as the destructive path below,
-        # against the in-memory checkpoint, writing nothing — INCLUDING the
-        # seed: a target superseded out of the live checkpoint (or ledger-
-        # only) never appears in the walk, but its own id still reaches the
-        # amendments keyed on it.
-        spliced = {str(target["id"] or "")}
-        if isinstance(checkpoint, dict):
-            for section, key in store._ITEM_LISTS:
-                for i in (checkpoint.get(section) or {}).get(key) or []:
-                    if isinstance(i, dict) and (
-                            i.get("id") == target["id"]
-                            or normalize.content_key(i.get("text") or "")
-                            == value_key):
-                        spliced.add(str(i.get("id") or ""))
-        spliced.discard("")
-        ref_reach = sorted({str(row.get("refutation_id") or "")
-                            for row in refutations.events(project_dir=project)
-                            if value_key in refutations.row_content_keys(row)})
-        amend_reach = sorted({
-            str(row.get("amendment_id") or "")
-            for row in amendments.events(project_dir=project)
-            if value_key in amendments.row_content_keys(row)
-            or str(row.get("item_id") or "") in spliced})
-        print(f"would forget {target['id']}: {target.get('text', '')}")
-        active_rulings = sorted(
-            rid for rid in set(ref_reach) | {str(target["id"])}
-            if ledger_meta.get(rid) == ("ruling", "active"))
-        if active_rulings:
-            print("WARNING — this removes ACTIVE ruling(s): "
-                  + ", ".join(active_rulings))
-        sibling_ids = sorted(spliced - {str(target["id"])})
-        if sibling_ids:
-            print("also removed — checkpoint siblings holding the value: "
-                  + ", ".join(sibling_ids))
-        also = [rid for rid in ref_reach + amend_reach
-                if rid and rid != target["id"]]
-        if also:
-            print("also removed — ledger records reached by value or by "
-                  "doomed item id: " + ", ".join(also))
-        return 0
-    # #402: key the tombstone on the CANONICAL value (normalize.content_key),
-    # not the raw bytes — so a later re-extraction of the same claim (different
-    # case, invisible chars, a look-alike glyph) folds to the same key and is
-    # suppressed at capture. Still a hash, never the text: removal means the
-    # content leaves the audit trail too (#321).
-    content_hash = normalize.content_key(target.get("text") or "")
-    # #693: forget takes no --by and asks no confirmation, so without this
-    # gate it is an agent path that un-renders a human-ratified standing
-    # constraint — the one thing no agent path may do. A human at a terminal
-    # proceeds (deletion stays the user's call, #421 posture unchanged); a
-    # non-interactive caller is pointed at the human.
-    doomed_rulings = sorted(
-        rid for rid in ({
-            str(row.get("refutation_id") or "")
-            for row in refutations.events(project_dir=project)
-            if content_hash in refutations.row_content_keys(row)}
-            | {str(target["id"])})
-        if ledger_meta.get(rid) == ("ruling", "active"))
-    if doomed_rulings:
-        if not sys.stdin.isatty():
-            print("refused: this would remove ACTIVE ruling(s) "
-                  + ", ".join(doomed_rulings)
-                  + " — a human decision. Run it from a terminal, or ask the "
-                  "user; `daimon ruling retire` records the verdict instead.")
-            return 1
-        # Deleting what renders is strictly more power than rewriting it,
-        # and rewriting confirms. A human can still say y (#421 unchanged).
-        print("WARNING — this removes ACTIVE ruling(s): "
-              + ", ".join(doomed_rulings))
-        answer = input("Remove? [y/N]: ").strip().casefold()
-        if answer not in ("y", "yes"):
-            print("not removed")
-            return 1
-    sid = str((checkpoint or {}).get("session_id") or "")
-    # A checkpoint-bearing project still needs its session id to rewrite. A
-    # ledger-only value has no checkpoint to rewrite, so the missing id is not
-    # an error there.
-    if isinstance(checkpoint, dict) and not sid:
-        print("checkpoint has no session_id — cannot rewrite")
-        return 1
-    # Tombstone BEFORE the rewrite (#418): write_checkpoint's forget gate
-    # consults the ledger during the write — appended after, the new key is
-    # invisible to that scrub and sibling ids carrying the same value survive.
-    # Failing here leaves the checkpoint untouched: no half-removal without an
-    # audit-trail record. allow_disabled (#421): forget is the ratified
-    # deletion exemption to the kill switch — the tombstone (and the rewrite
-    # below, which #418 chains to it) must land even while daimon is disabled.
-    ok = store.append_event(target["id"], f"forgotten:{content_hash}",
-                            note=args.reason or "", kind="tombstone",
-                            project_dir=project, allow_disabled=True)
-    if not ok:
-        print("tombstone event not written (project unknown or ledger unwritable)")
-        return 1
-    # Splice by VALUE, not only id (#418): one value can hold sibling ids —
-    # the same sentence in two sections, or a widened hash within one
-    # (store._stamp_item_ids). Removal is content removal, so every item
-    # folding to the tombstoned key goes. Id kept in the predicate as a belt
-    # for non-string text, which content_key canonicalizes to "".
-    spliced_ids = {str(target["id"] or "")}
-    if isinstance(checkpoint, dict):
-        for section, key in store._ITEM_LISTS:
-            lst = (checkpoint.get(section) or {}).get(key)
-            if isinstance(lst, list):
-                doomed = [i for i in lst
-                          if isinstance(i, dict)
-                          and (i.get("id") == target["id"]
-                               or normalize.content_key(i.get("text") or "")
-                               == content_hash)]
-                # #691: every spliced sibling id, not only the named target —
-                # amendments are keyed by item id, and an amendment about a
-                # sibling holding the same value is plaintext ABOUT the
-                # forgotten content (the #418 sibling rule, extended to the
-                # ledger that references the siblings).
-                spliced_ids.update(str(i.get("id") or "") for i in doomed)
-                lst[:] = [i for i in lst if i not in doomed]
-        # allow_disabled (#421): the ONE write_checkpoint call that may run under
-        # the kill switch — the rewrite that makes the deletion real on disk.
-        # rotate=False: rotation copies the CURRENT latest into prev-1 before
-        # writing, and the current latest is the PRE-forget bytes. Rotating here
-        # made the deletion manufacture a fresh copy of the value it was asked to
-        # remove, in a file that did not exist when the user ran the command.
-        store.write_checkpoint(sid, checkpoint, project_dir=project,
-                               allow_disabled=True, rotate=False)
-    # The live checkpoint is one surface of several. prev-N and superseded
-    # session files hold the same plaintext and were never in the contract
-    # (#419: plaintext is what puts a file inside it, not its role). Runs even
-    # with no live checkpoint: a ledger-only project still has these surfaces.
-    store.scrub_content_key(content_hash, project_dir=project)
-    # #600 slice A: the author's own team-mirror copies are plaintext this
-    # machine owns (#419) — scrubbed here; teammates' copies and upstream
-    # git history are the sync protocol's to converge (tombstone
-    # propagation), not a local rewrite's.
-    team_scrubbed = store.scrub_team_copies(content_hash,
-                                            project_dir=project)
-    # #600 slice B: publish the deletion itself (hash only) so teammates can
-    # suppress the value without waiting to pull the scrubbed file — and so
-    # a copy THEY extracted independently can be acted on at all.
-    store.publish_tombstone(content_hash, project_dir=project)
-    # #599: rows appended BEFORE this forget can carry the value in
-    # `item_text`/`status`/`note` — redacted in place, rows never dropped
-    # (the one ratified rewrite of the append-only ledger).
-    events_scrubbed = store.scrub_event_fields(content_hash,
-                                               project_dir=project)
-    # #578: same value, second plaintext store. The ledger splices on the SAME
-    # canonical key for the same reason the checkpoint splices on it rather than
-    # on the id — removal is content removal, so a refutation asserting the
-    # forgotten value goes with it whether or not it was the named target.
-    forgotten_refutations = refutations.forget_content_key(
-        content_hash, project_dir=project)
-    # #678 fork A: relation rows hold no text, but an edge touching this item
-    # is an equivalence CLAIM about its content (an `exact-text` rail against
-    # a surviving twin re-derives what the value was), and post-forget the
-    # relations ledger would be the only surface binding the forgotten id to
-    # its sessions, kind, and revision-chain length. The scrub is id-keyed —
-    # the tombstone above landed on this exact id — and the audit's
-    # relations-ledger scan is what proves it reached the edges.
-    forgotten_relations = relations.forget_item_id(
-        target["id"], project_dir=project)
-    # #691: same value, another plaintext store — and unlike relations, amend
-    # rows DO carry prose, so records targeting a forgotten item (or any of
-    # its spliced siblings) go with it — their evidence may paraphrase the
-    # removed content — and records holding the value in any plaintext field
-    # go regardless of target.
-    spliced_ids.discard("")
-    forgotten_amendment_set = set(
-        amendments.forget_content_key(content_hash, project_dir=project))
-    for doomed_id in sorted(spliced_ids):
-        forgotten_amendment_set.update(
-            amendments.forget_item_id(doomed_id, project_dir=project))
-    forgotten_amendments = sorted(forgotten_amendment_set)
-    # #422: the serializer chunk cache holds PRE-redaction extraction output
-    # (quote verification forbids redacting before caching, #125), keyed by
-    # chunk text — the forgotten value cannot be located selectively, so the
-    # purge is WHOLESALE and default-on. Never fatal: the belief-state
-    # deletion above is the primary contract; a failed purge is reported
-    # honestly below, and the age reaper still bounds any survivor at
-    # chunk_cache_days.
-    try:
-        purged, purge_err = serializer.purge_chunk_cache()
-    except Exception as e:  # belt: purge_chunk_cache itself never raises
-        purged, purge_err = 0, str(e)
-    # #607: the Windsurf adapter writes its own transcripts when Cascade
-    # gives it none — daimon-authored plaintext, so inside the contract.
-    # Wholesale for the same reason as the chunk cache: the tombstone is a
-    # hash, so a value inside prose cannot be located to remove selectively.
-    try:
-        ws_purged, ws_err = store.purge_windsurf_state()
-    except Exception as e:  # belt: purge_windsurf_state never raises
-        ws_purged, ws_err = 0, str(e)
-    # #605: serialize-crash.log is the detached child's RAW stderr — an
-    # uncaught traceback carries whatever the crashing frame held, and the
-    # bytes were never scrubbed (status redacted its tail on READ, #513, over
-    # a file nothing deleted). Wholesale like the two purges above: the
-    # tombstone is a hash, so a value inside a traceback cannot be located.
-    try:
-        crash_purged, crash_err = store.purge_crash_log()
-    except Exception as e:  # belt: purge_crash_log never raises
-        crash_purged, crash_err = 0, str(e)
-    # #616: backend-stderr.log holds backend stderr/stdout, which CLI
-    # backends can seed with transcript text (#141). Wholesale like the
-    # crash sink: prose diagnostics, hash tombstone, value unlocatable.
-    try:
-        backend_purged, backend_err = store.purge_backend_stderr_log()
-    except Exception as e:  # belt: purge_backend_stderr_log never raises
-        backend_purged, backend_err = 0, str(e)
-    # #616: pre-fix downgrade lines logged item text into serialize.log.
-    # Shape-targeted scrub, NOT a purge — serialize.log is also the ledger
-    # `status` parses, and the capture record must survive a forget.
-    try:
-        scrubbed_lines, scrub_err = store.scrub_serialize_log()
-    except Exception as e:  # belt: scrub_serialize_log never raises
-        scrubbed_lines, scrub_err = 0, str(e)
-    _note_usage("forget")
-    surfaces = []
-    if isinstance(checkpoint, dict):
-        surfaces.append("the live checkpoint")
-    if forgotten_refutations:
-        gone_rulings = [r for r in forgotten_refutations
-                        if ledger_meta.get(r, ("refutation", None))[0]
-                        == "ruling"]
-        gone_refuts = [r for r in forgotten_refutations
-                       if r not in gone_rulings]
-        if gone_refuts:
-            surfaces.append(f"{len(gone_refuts)} refutation(s) "
-                            f"({', '.join(gone_refuts)})")
-        if gone_rulings:
-            surfaces.append(f"{len(gone_rulings)} ruling(s) "
-                            f"({', '.join(gone_rulings)})")
-    if forgotten_relations:
-        surfaces.append(f"{len(forgotten_relations)} relation(s) "
-                        f"({', '.join(forgotten_relations)})")
-    if forgotten_amendments:
-        surfaces.append(f"{len(forgotten_amendments)} amendment(s) "
-                        f"({', '.join(forgotten_amendments)})")
-    print(f"forgot {target['id']} (content hash {content_hash}) — "
-          f"removed from {' and '.join(surfaces) or 'no store'}; "
-          "tombstone recorded")
-    if team_scrubbed:
-        print(f"scrubbed {len(team_scrubbed)} own team-mirror cop(y/ies) — "
-              "run `daimon team sync` to publish; teammates' copies and "
-              "upstream git history remain until tombstone propagation")
-    if events_scrubbed:
-        print(f"redacted {events_scrubbed} event-ledger field(s) "
-              "carrying the value (rows kept, field replaced)")
-    if purge_err is not None:
-        print(f"warning: chunk cache purge failed: {purge_err} — "
-              "cached pre-redaction chunks may persist up to "
-              f"{config.chunk_cache_days()} day(s) (age reaper)")
-    else:
-        print(f"purged {purged} cached chunk extraction(s) "
-              "(pre-redaction serializer cache)")
-    if ws_err is not None:
-        print(f"warning: windsurf transcript purge failed: {ws_err} — "
-              "daimon-authored conversation text may persist up to "
-              f"{config.windsurf_state_days()} day(s) (age reaper)")
-    else:
-        # Always printed, like the chunk-cache line: a silent zero is how a
-        # misconfigured store (writer and deleter disagreeing about the
-        # directory) stays invisible. The scope note is not decoration —
-        # the store is keyed by trajectory and carries no project
-        # attribution, so this purge is machine-wide by construction.
-        print(f"purged {ws_purged} daimon-authored windsurf transcript "
-              "file(s) across all projects (machine-wide: the store is keyed "
-              "by trajectory, not project); host-authored transcripts are "
-              "untouched")
-    if crash_err is not None:
-        print(f"warning: crash log purge failed: {crash_err} — serializer "
-              "tracebacks may persist (bounded to a trimmed tail at the next "
-              "spawn, never removed)")
-    else:
-        # Printed even at zero, like the two lines above: a silent zero is
-        # how a writer and a deleter disagreeing about DAIMON_LOG_DIR stays
-        # invisible. One crash log per machine, so the scope note is the
-        # same honesty the windsurf line owes.
-        print(f"purged {crash_purged} serializer crash log(s) across all "
-              "projects (machine-wide: the log is raw child stderr, not "
-              "keyed by project)")
-    if backend_err is not None:
-        print(f"warning: backend log purge failed: {backend_err} — backend "
-              "echo of transcript text may persist (byte-bounded at the "
-              "write seam, never removed)")
-    else:
-        # Same silent-zero rule as the three lines above.
-        print(f"purged {backend_purged} backend stderr log(s) across all "
-              "projects (machine-wide: backend diagnostics are not keyed "
-              "by project)")
-    if scrub_err is not None:
-        print(f"warning: serialize.log scrub failed: {scrub_err} — legacy "
-              "downgrade lines may still carry item text")
-    else:
-        # Printed even at zero, same rule again: a scrubber resolving a
-        # different log dir than the writer must not read as "nothing left".
-        print(f"scrubbed {scrubbed_lines} legacy downgrade line(s) in "
-              "serialize.log (payload replaced; ledger lines kept)")
-    return 0
-
-
-def _is_supersede_candidate(item_id: str, project) -> bool:
-    """True when the item's LATEST lifecycle event is a serializer-authored
-    supersede-candidate — an unconfirmed machine SUGGESTION (#14) that has
-    never withheld anything. This is the one reopen target that needs no
-    evidence: rejecting a guess is not vouching for a claim (#111). The
-    source gate mirrors `_emit_supersede_candidates` — only the serializer
-    ever writes candidates, so a human verdict already latest reads as
-    not-a-candidate and keeps the evidence gate. Fails closed: a broken
-    events fold means 'not a candidate', so the evidence gate still applies."""
-    try:
-        prior = store.resolutions(project_dir=project).get(item_id)
-    except Exception:
-        return False
-    if not isinstance(prior, dict):
-        return False
-    status = str(prior.get("status") or "").lower()
-    source = str(prior.get("source") or "")
-    return status.startswith("supersede-candidate") and source == "serializer"
-
-
-def _cmd_reverify(args) -> int:
-    """Evidence-gated reopen (#103): re-stamping a resolved item without
-    evidence would mark an unchecked claim verified — the one thing this
-    tool must never do to its own audit trail. Reopen is allowed only when
-    the original anchor still checks out live (the code proved itself) or
-    the caller supplies --evidence (a human vouches for it); otherwise the
-    refusal appends nothing. Exact id only — no fuzzy match, because
-    reopening is a deliberate act; find ids via `daimon status --suppressed`.
-
-    #111 exception: when the target is an unconfirmed supersede CANDIDATE
-    (a machine guess that has never withheld anything), reopen is allowed
-    with no evidence — rejecting a suggestion is not vouching for a claim.
-    The reopened event is a human verdict, so re-detection stays silent."""
-    project = _resolve_project(args.project)
-    checkpoint = store.read_latest(project_dir=project, fallback=False)
-    if not isinstance(checkpoint, dict):
-        print("no checkpoint for this project yet — nothing to reverify")
-        return 1
-    item = None
-    for section, key in store._ITEM_LISTS:
-        for it in ((checkpoint.get(section) or {}).get(key) or []):
-            if isinstance(it, dict) and it.get("id") == args.target:
-                item = it
-                break
-        if item is not None:
-            break
-    if item is None:
-        print(f"no item found with id {args.target!r}")
-        return 1
-    evidence = args.evidence or ""
-    a = item.get("anchored_to")
-    if isinstance(a, dict) and anchor.check(a, project) == "live":
-        note = "reverified: anchor live"
-        if evidence:
-            note += "; " + evidence
-    elif evidence:
-        note = f"evidence: {evidence}"
-    elif _is_supersede_candidate(item["id"], project):
-        # #111: the target is an unconfirmed machine SUGGESTION, never a
-        # withheld/suppressed item — rejecting a guess needs no proof. The
-        # reopened event below is a human verdict, so the human-speaks-once
-        # gate (#14) silences re-detection of this candidate permanently.
-        note = "candidate rejected"
-    else:
-        print("re-stamping without evidence would mark an unchecked claim "
-              "verified — supply --evidence, or fix the anchored code and retry")
-        return 1
-    ok = store.append_event(item["id"], "reopened", note=note,
-                            item_text=item.get("text", ""), project_dir=project)
-    if not ok:
-        print("event not written (daimon disabled or project unknown)")
-        return 1
-    # A typed baseline for this verb. Without it reverify's count is
-    # structurally zero, and a zero that measures missing instrumentation
-    # cannot be compared against a UI channel's count later.
-    _note_usage("reverify")
-    print(f"reopened {item['id']}: {item.get('text', '')}")
-    return 0
-
-
-def _cmd_loops(args) -> int:
-    """`daimon loops` (#480 slice 1): list open, briefable loop items with
-    ids for this project — the read-only counterpart to the id handles
-    briefing._line now renders inline (an agent, or a human, needs something
-    to pass to `daimon resolve`).
-
-    Reuses the SAME item walk `_cmd_resolve` builds (store._ITEM_LISTS over
-    the latest checkpoint) and the SAME withhold classification `daimon
-    status --suppressed` uses (briefing.withhold over store.resolutions) —
-    the resolved/live split must stay in exactly one place, or this listing
-    could show an item the briefing itself would withhold (an agent would
-    then "resolve" a ghost).
-
-    Briefable = briefing.BRIEFABLE_ITEM_KEYS (open_questions — external and
-    non-external both — plus uncertainties). Decisions/beliefs/
-    contradictions are valid `daimon resolve` targets too, but are not
-    loop-shaped; listing them here would invite resolving settled facts
-    (#480 scope guard, mirrors briefing._line's new suffix)."""
-    _note_usage("loops")
-    project = _resolve_project(args.project)
-    checkpoint = store.read_latest(project_dir=project, fallback=False)
-    if not isinstance(checkpoint, dict):
-        print("no checkpoint for this project yet — nothing to list")
-        return 0
-    try:
-        events = store.resolutions(project_dir=project)
-        # #691: amendments ride along so the listing agrees with the
-        # briefing — an agent discovering targets here must see that an
-        # item is already amended, or its cheapest probe is a duplicate
-        # proposal.
-        checkpoint, _, _ = briefing.withhold(
-            checkpoint, events,
-            amendments=amendments.renderable(project_dir=project))
-    except Exception:
-        pass  # fail-open, same stance as _print_suppressed
-    rows = []
-    for section, key in store._ITEM_LISTS:
-        if key not in briefing.BRIEFABLE_ITEM_KEYS:
-            continue
-        for item in ((checkpoint.get(section) or {}).get(key) or []):
-            if not isinstance(item, dict) or not item.get("id"):
-                continue
-            text = str(item.get("text") or "").strip()
-            if not text:
-                continue
-            if item.get("_agent_claim"):
-                # #480 slice 4: the listing agrees with the briefing — an
-                # item carrying a still-pending, unverified agent claim
-                # (briefing.withhold's transient stamp) is marked here too.
-                text += " (agent claim pending)"
-            amend_stamp = item.get("_amend")
-            if isinstance(amend_stamp, dict):
-                n = (len(amend_stamp.get("rows") or [])
-                     + (amend_stamp.get("overflow") or 0))
-                if n:
-                    text += f" (amended ×{n})"
-            rows.append((item["id"], key, text, briefing._mark(item)))
-    if not rows:
-        print("no open loops")
-        return 0
-    for item_id, key, text, mark in rows:
-        print(f"  {item_id}  [{key}] [{mark}] {text}")
     return 0
 
 
@@ -3204,214 +1888,8 @@ def _cmd_status(args) -> int:
 def _cmd_mcp_serve(args) -> int:
     """#261: blocking stdio MCP server. No usage note here — each tool call
     notes `mcp:<tool>` itself; serving is not reading."""
-    from . import mcp_server
+    from .. import mcp_server
     return mcp_server.serve()
-
-
-def _load_audit_transcript(tpath):
-    """Parse one transcript file into the (haystack, texts_by_id) pair
-    audit-quotes checks quotes against — the one place #503's per-session
-    cache calls into transcript.from_file, so it is charged once per session
-    no matter how many items or checkpoints resolve to it."""
-    msgs = transcript.from_file(tpath)
-    # #440: the same stripped haystacks verify_quotes used at serialize time.
-    # Audit re-blesses stored items, so reading the raw render here would
-    # certify exactly the echo verification rejected — and do it with the
-    # CLI's authority. #512: daimon-output tool rows blank here too, for the
-    # same reason and by the same rule.
-    haystack = serializer.stripped_transcript(msgs)
-    daimon_ids = serializer.daimon_output_ids(msgs)
-    texts_by_id = {
-        mid: "" if mid in daimon_ids else serializer.strip_injected(text)
-        for mid, text in serializer.message_texts_by_id(msgs).items()}
-    return haystack, texts_by_id
-
-
-def _legacy_audit_source(session_id, author=None):
-    """Explicitly inferred pre-#594 Claude candidate, never a bound receipt."""
-    if not provenance.valid_session_id(session_id):
-        return None
-    source = {
-        "version": provenance.SOURCE_REF_VERSION,
-        "host": "claude-code",
-        "session_id": session_id,
-        "locator": "managed",
-    }
-    if isinstance(author, str) and author.strip():
-        source["author"] = author.strip()
-    return source
-
-
-def _audit_item_source(item):
-    """Return (source, receipt) without containing-checkpoint fallback."""
-    receipt = item.get("quote_provenance")
-    if provenance.valid_quote_receipt(receipt):
-        return receipt["source"], receipt
-    origin = item.get("origin_session")
-    if not provenance.valid_session_id(origin):
-        return None, None
-    origin_cp = store.read_checkpoint(origin)
-    if isinstance(origin_cp, dict):
-        source = origin_cp.get("source_ref")
-        if provenance.valid_source_ref(source):
-            return source, None
-    return _legacy_audit_source(origin, item.get("origin_author")), None
-
-
-def _resolve_audit_source(source, resolver, cache: dict):
-    """Resolve and parse one strict source once per complete source identity."""
-    if not provenance.valid_source_ref(source):
-        return None
-    key = json.dumps(source, sort_keys=True, separators=(",", ":"))
-    if key in cache:
-        return cache[key]
-    result = None
-    resolved = resolver.resolve(source)
-    if resolved.state == "resolved" and resolved.path is not None:
-        try:
-            result = _load_audit_transcript(resolved.path)
-        except (OSError, FileNotFoundError):
-            result = None
-    cache[key] = result
-    return result
-
-
-def _cmd_audit_privacy(args) -> int:
-    """Read-only tombstone residue audit — proves forget's contract instead
-    of trusting it (#583: a passing test once asserted the residue). Exit 0
-    proven clean / 1 residue / 3 cannot-prove; 3 exists because "could not
-    check" must never look like "all clean"."""
-    if getattr(args, "all_projects", False):
-        results = privacy.audit_all()
-    else:
-        results = [privacy.audit_project(_resolve_project(args.project))]
-    render.render_privacy_audit(results)
-    code = privacy.exit_code(results)
-    # One tag per OUTCOME: "the auditor ran" and "the auditor found residue"
-    # answer different questions, and folding them loses the only number that
-    # says whether the deletion contract holds in the field.
-    _note_usage({1: "audit-privacy:residue",
-                 3: "audit-privacy:unproven"}.get(code, "audit-privacy"))
-    return code
-
-
-def _cmd_audit_quotes_deprecated(args) -> int:
-    print("note: 'daimon audit-quotes' is deprecated — use 'daimon audit quotes'")
-    return _cmd_audit_quotes(args)
-
-
-def _cmd_audit_quotes(args) -> int:
-    """Read-only audit (#125): re-check every stored verbatim quote against its
-    source transcript with the SAME tier-f matcher serialize uses, and REPORT.
-
-    #594: a valid item receipt is authoritative for source identity and message
-    binding. Legacy origin_session may form an explicitly inferred Claude
-    candidate, but an absent/unresolved source NEVER falls back to the
-    containing checkpoint: exact-text carry proves origin, quote evidence, and
-    containing session can be three different facts.
-
-    Never rewrites a trust tag — a blind backfill would flip a large share of
-    the historical corpus. Measured (#503): the overwhelming majority of
-    failures on an unpatched checker were exactly this resolution bug, not
-    quotes crossing content the current renderer no longer emits — that class
-    is a small minority. Default scope is the current project; --all spans
-    the whole corpus."""
-    project = _resolve_project(args.project)
-    want_slug = store.project_slug(project)
-    d = config.checkpoint_dir()
-    try:
-        files = store._session_files(d)
-    except OSError:
-        files = []
-    scanned = paired = unpaired = items = verified = failed = id_resolved = 0
-    origin_resolved = 0
-    transcripts: dict = {}
-    resolver = provenance.SourceResolver(
-        claude_projects=config.claude_projects_dir(),
-        current_author=config.author())
-    failures: list[tuple[str, str]] = []
-    for f in sorted(files):
-        try:
-            cp = json.loads(f.read_text(encoding="utf-8"))
-        except (OSError, json.JSONDecodeError):
-            continue
-        if not isinstance(cp, dict):
-            continue
-        slug = cp.get("project_slug")
-        if not args.all and want_slug is not None and slug != want_slug:
-            continue
-        scanned += 1
-        session_id = str(cp.get("session_id") or f.stem)
-        own_source = cp.get("source_ref")
-        if not provenance.valid_source_ref(own_source):
-            own_source = _legacy_audit_source(session_id, cp.get("author"))
-        own = _resolve_audit_source(own_source, resolver, transcripts)
-        if own is not None:
-            paired += 1
-        else:
-            unpaired += 1
-        for item in serializer.iter_items(cp):
-            if item.get("trust") != "verbatim":
-                continue
-            quote = item.get("quote")
-            if not isinstance(quote, str) or not quote.strip():
-                continue
-            source, receipt = _audit_item_source(item)
-            resolved = _resolve_audit_source(source, resolver, transcripts)
-            if resolved is None:
-                continue
-            haystack, texts_by_id = resolved
-            items += 1
-            if source.get("session_id") != session_id:
-                origin_resolved += 1
-            # #358: an item bound to source message id(s) resolves the id and
-            # compares bytes against just that message. Missing/invalid ids
-            # (old checkpoints, moved/truncated transcripts) fall back to the
-            # whole-transcript scan — the pre-#358 verdict, byte-identical.
-            checked_item = item
-            if receipt is not None:
-                checked_item = dict(item)
-                ids = provenance.binding_message_ids(receipt)
-                if ids:
-                    checked_item[serializer.SOURCE_IDS_KEY] = ids
-                else:
-                    checked_item.pop(serializer.SOURCE_IDS_KEY, None)
-            scoped = serializer.scoped_haystack(checked_item, texts_by_id)
-            if scoped is not None and serializer.quote_matches(quote, scoped):
-                verified += 1
-                id_resolved += 1
-            elif serializer.quote_matches(quote, haystack):
-                verified += 1
-            else:
-                failed += 1
-                failures.append((session_id, str(item.get("text") or "")))
-    rate = (verified / items) if items else 0.0
-    scope = "all projects" if args.all else project
-    lines = [
-        f"audit-quotes ({scope})",
-        f"  checkpoints scanned: {scanned}  paired: {paired}  unpaired: {unpaired}",
-        f"  verbatim quotes checked: {items}  verified: {verified}  "
-        f"failed: {failed}  id-resolved: {id_resolved}  "
-        f"origin-resolved: {origin_resolved}  rate: {rate:.1%}",
-    ]
-    if failures:
-        top = max(0, args.top)
-        lines.append(f"  top {min(top, len(failures))} failures (item text prefix):")
-        for sid, text in failures[:top]:
-            lines.append(f"    [{sid}] {text[:80]}")
-    print("\n".join(lines))
-    # #504: the only read-side verification verb recorded nothing, so there was
-    # no evidence either way about whether anyone reaches for it. The unpaired
-    # variant is a distinct event, not a detail of this one: a run that resolved
-    # no transcript verified nothing, and it is also how a host whose
-    # transcripts live outside the registered resolver's reach shows up at all.
-    # #503: keyed on ANY resolved transcript, not on `paired`. Once resolution
-    # is per item, a checkpoint whose own transcript is gone still verifies its
-    # carried items through origin_session — `paired` counts containing
-    # checkpoints and would report that run as silence.
-    resolved_any = any(v is not None for v in transcripts.values())
-    _note_usage("audit-quotes" if resolved_any else "audit-quotes:unpaired")
-    return 0
 
 
 def _cmd_heal(args) -> int:
@@ -3478,134 +1956,6 @@ def _cmd_heal(args) -> int:
 
 
 # ---- team: sidecar private-repo sync (#113) ----
-
-
-def _cmd_team_init(args) -> int:
-    try:
-        # #279: a fresh team is born scoped to the project init ran from.
-        dest = teamsync.init(args.remote_url, project_dir=Path.cwd())
-    except teamsync.TeamError as exc:
-        print(f"error: {exc}", file=sys.stderr)
-        return 1
-    render.render_team_init([
-        f"initialized team sidecar: {dest}",
-        "checkpoints now sync there — `daimon team sync` runs opportunistically "
-        "at session start",
-    ])
-    return 0
-
-
-def _cmd_team_sync(args) -> int:
-    """rc 0 for every sync-nothing-to-do shape (no git, no remotes, offline);
-    warnings go to stderr but never change the rc — a degraded sync is not a
-    user error."""
-    if getattr(args, "project", None):
-        # Accepted for CLI symmetry only — say so instead of silently running
-        # a global sync the user thought was scoped (#29).
-        print("daimon team: --project is ignored — sync is project-agnostic "
-              "(all own checkpoints sync)", file=sys.stderr)
-    if not teamsync.git_available():
-        render.render_team_sync(["daimon team: git not found on PATH — sync skipped"])
-        return 0
-    reports = teamsync.sync()
-    # #600 slice B, opt-in: apply teammates' tombstones to THIS machine's own
-    # checkpoints. TWO gates, and the flag is the load-bearing one: bare
-    # `daimon team sync` is spawned DETACHED at SessionStart by
-    # lib.spawn_team_sync with stdout to DEVNULL, exactly like heal — so a
-    # setting alone would delete local belief state unattended and silently,
-    # which is the failure this design exists to prevent. The hook never
-    # passes --apply-forget, so only a typed command can reach this.
-    # Machine-wide, matching sync's own project-agnostic contract.
-    if getattr(args, "apply_forget", False):
-        if not config.team_apply_forget():
-            print("daimon team: --apply-forget needs DAIMON_TEAM_APPLY_FORGET=1"
-                  " — a teammate's forget rewriting your own checkpoints is"
-                  " opt-in, and there is no undo", file=sys.stderr)
-        else:
-            applied = store.apply_foreign_tombstones(all_projects=True)
-            print(f"applied teammates' forget tombstones to {len(applied)} "
-                  "local surface(s) across all projects")
-    # #246: fetched teammate files are fingerprint input — freshen here (the
-    # SessionStart hook spawns sync detached, off the prompt path) so the
-    # first recall after a fetch doesn't pay the rebuild. Unconditional on
-    # purpose: a no-op sync warms in ~ms, and any staleness left by other
-    # writers gets healed opportunistically.
-    recall.warm()
-    if not reports:
-        render.render_team_sync([
-            "daimon team: no team remote configured — nothing to sync "
-            "(run `daimon team init <remote-url>`)",
-        ])
-        return 0
-    for r in reports:
-        parts = [f"{r['committed']} committed", "pushed" if r["pushed"] else "no push"]
-        if r["fetched"]:
-            parts.append("fetched teammates' updates")
-        line = f"{r['slug']}: " + ", ".join(parts)
-        if r["notes"]:
-            line += " (" + "; ".join(r["notes"]) + ")"
-        # Rendered per-report (not collected and rendered once after the loop):
-        # this keeps a report's stdout line interleaved with ITS OWN stderr
-        # warnings below, matching the ordering the pre-#68 print()-per-report
-        # loop had.
-        render.render_team_sync([line])
-        for w in r["warnings"]:
-            print(f"warning: {w}", file=sys.stderr)
-    return 0
-
-
-def _cmd_team_status(args) -> int:
-    if not teamsync.git_available():
-        render.render_team_status(["daimon team: git not found on PATH"])
-        return 0
-    rows = teamsync.team_status()
-    if not rows:
-        render.render_team_status([
-            "no team remote configured — run `daimon team init <remote-url>`",
-        ])
-        return 0
-    if args.json:
-        print(json.dumps(rows, indent=2))
-        return 0
-    lines = []
-    for row in rows:
-        authors = ", ".join(row["authors"]) or "none yet"
-        # #217: uncommitted checkpoints (dual-written to disk, staged+committed
-        # only on the next sync) are invisible to `unpushed` alone — surface
-        # them and nudge a sync, but only when there's something to nudge
-        # about, so the common case's wording stays byte-identical.
-        pending = row.get("pending", 0)
-        pending_part = (f", {pending} pending checkpoint(s) — run "
-                        "`daimon team sync`") if pending > 0 else ""
-        lines.append(f"{row['slug']}: {row['freshness']} — "
-                     f"{row['unpushed']} unpushed checkpoint(s)"
-                     f"{pending_part}, authors: {authors}")
-        # #200: a broken daimon-team.toml fails open (mapping ignored) on the
-        # write path — status is the one place the parse error surfaces.
-        if row.get("config_warning"):
-            lines.append(f"  warning: {row['config_warning']}")
-        # #279: scope is default-closed — an empty allowlist means the remote
-        # receives nothing, which must be visible, not a silent surprise.
-        scope = row.get("scope") or []
-        if scope:
-            lines.append(f"  scope: {', '.join(scope)}")
-        elif config.team_project():
-            lines.append("  scope: none configured — DAIMON_TEAM_PROJECT "
-                         "grants this machine's sessions")
-        else:
-            lines.append("  scope: none — this remote receives no checkpoints "
-                         "(add [scope] repos to daimon-team.toml)")
-    # #387: where the CURRENT project's checkpoints route — the one place a
-    # misconfigured toml is visible BEFORE a session's checkpoint quietly
-    # stays in the local mirror.
-    dests = store._team_write_slugs(_resolve_project(getattr(args, "project", None)))
-    line = "this project writes to: " + ", ".join(dests)
-    if dests == [store._TEAM_LOCAL_REMOTE]:
-        line += ("  (no remote grants it membership — add its repo URL to a "
-                 "sidecar's daimon-team.toml [scope] repos)")
-    lines.append(line)
-    render.render_team_status(lines)
-    return 0
 
 
 # ---- configure: detect/report the resolved backend + fill gaps in ~/.daimon/env ----
@@ -4188,57 +2538,6 @@ def _host_scripts(spec) -> str:
                      if n not in ("_daimon_hook_lib.py", "redact.py"))
 
 
-def _cmd_hooks_list(args) -> int:
-    lines = [f"{host}  ({_host_scripts(spec)}; events: {', '.join(spec['events'])})"
-             for host, spec in sorted(_HOOK_HOSTS.items())]
-    render.render_hooks_list(lines)
-    return 0
-
-
-def _cmd_hooks_install(args) -> int:
-    """Copy the host's packaged hook script(s) to ~/.daimon/hooks/ — a STABLE
-    path the host's hooks config points at once. Idempotent: re-running after
-    `uv tool upgrade daimon-briefing` refreshes the scripts to match the
-    installed CLI, which is the whole point (#43: a curl'd script drifts)."""
-    from importlib import resources
-
-    spec = _HOOK_HOSTS.get(args.host)
-    if spec is None:
-        known = ", ".join(sorted(_HOOK_HOSTS))
-        print(f"error: unknown host '{args.host}' (known: {known})", file=sys.stderr)
-        return 2
-    pkg = resources.files("daimon_briefing._hooks")
-    if spec.get("register") == "codex":
-        # Codex owns its own install path: two scripts registered under two
-        # events straight into ~/.codex/hooks.json (#262), not a printed snippet.
-        from . import codex_hooks
-
-        render.render_hooks_install(codex_hooks.install(pkg, Path.home()))
-        return 0
-    target = _hooks_target_dir()
-    target.mkdir(parents=True, exist_ok=True)
-    for name in spec["files"]:
-        data = (pkg / name).read_bytes()
-        dest = target / name
-        dest.write_bytes(data)
-        dest.chmod(dest.stat().st_mode | 0o100)  # u+x
-    entry = target / spec["entry"]
-    lines = [
-        f"installed {len(spec['files'])} file(s) to {target}",
-        "",
-        "Register this command for the events below "
-        "(host hooks config — see the host's hooks documentation):",
-        f"  command: python3 {entry}",
-    ]
-    for ev in spec["events"]:
-        lines.append(f"  event:   {ev}")
-    lines.append("")
-    lines.append("Re-run `daimon hooks install " + args.host +
-                 "` after every `uv tool upgrade daimon-briefing`.")
-    render.render_hooks_install(lines)
-    return 0
-
-
 # ---- hooks status: audit installed copies against the packaged bytes (#266) ----
 #
 # A stale hook copy keeps *working* on old behavior after an upgrade, so drift is
@@ -4275,7 +2574,7 @@ def _codex_registration_status(home: Path) -> str:
     """REGISTERED / PARTIAL / UNREGISTERED for the ~/.codex/hooks.json entries.
     Reuses codex_hooks' own loader and ownership check so this verdict can never
     drift from what `hooks install codex` writes."""
-    from . import codex_hooks
+    from .. import codex_hooks
 
     settings = codex_hooks._load(home / ".codex" / "hooks.json")
     cfg = settings.get("hooks", {})
@@ -4401,72 +2700,6 @@ def _hook_drift_present() -> bool:
         return any(h["drift"] for h in _hooks_status_report(Path.home()))
     except Exception:
         return False
-
-
-def _cmd_hooks_status(args) -> int:
-    report = _hooks_status_report(Path.home())
-    if getattr(args, "json", False):
-        print(json.dumps(report, indent=2))
-    else:
-        render.render_hooks_status(report)
-    return 1 if any(h["drift"] for h in report) else 0
-
-
-# ---- skill: ship the portable agent skill from the package (#66) ----
-
-def _resolve_project_cwd() -> Path:
-    """--project writes relative to the repo root, not whatever subdirectory
-    the command was run from — same git-toplevel normalization _resolve_project
-    uses for checkpoint routing (#74); falls back to plain cwd outside a repo
-    or when git is unavailable (resolve_project_root's own contract)."""
-    return Path(config.resolve_project_root(str(Path.cwd())))
-
-
-def _cmd_skill_list(args) -> int:
-    from . import skill_install
-    rows = []
-    for host in sorted(skill_install.HOSTS):
-        scopes = [s for s in ("global", "project")
-                  if skill_install.HOSTS[host].get(s) is not None]
-        rows.append((host, scopes))
-    render.render_skill_list(rows)
-    return 0
-
-
-def _cmd_skill_show(args) -> int:
-    from . import skill_content
-    print(skill_content.render_compact() if args.compact
-          else skill_content.render_full(), end="")
-    return 0
-
-
-def _cmd_skill_install(args) -> int:
-    from . import skill_install
-    try:
-        lines = skill_install.install(
-            args.host, project=args.project, home=Path.home(),
-            cwd=_resolve_project_cwd())
-    except skill_install.SkillInstallError as exc:
-        print(f"error: {exc}", file=sys.stderr)
-        return 2
-    render.render_skill_lines(lines, footer=(
-        f"Re-run `daimon skill install {args.host}` after every "
-        "`uv tool upgrade daimon-briefing` to refresh the content.",
-    ))
-    return 0
-
-
-def _cmd_skill_uninstall(args) -> int:
-    from . import skill_install
-    try:
-        lines = skill_install.uninstall(
-            args.host, project=args.project, home=Path.home(),
-            cwd=_resolve_project_cwd())
-    except skill_install.SkillInstallError as exc:
-        print(f"error: {exc}", file=sys.stderr)
-        return 2
-    render.render_skill_lines(lines)
-    return 0
 
 
 def _crash_stamp_excepthook(exc_type, exc, tb) -> None:
@@ -4688,355 +2921,13 @@ def build_parser() -> argparse.ArgumentParser:
     )
     p_projects.set_defaults(func=_cmd_projects)
 
-    p_resolve = sub.add_parser(
-        "resolve", help="mark a checkpoint item resolved — append-only event, "
-        "folds at read so the item stops carrying (#102)",
-        epilog="Examples:\n  daimon resolve o-3f8a2c\n"
-               "  daimon resolve \"release pipeline approval\" --note \"shipped in 0.9\"\n",
-    )
-    p_resolve.add_argument("target", help="item id (exact) or a query that must match exactly one item")
-    p_resolve.add_argument("--status", default="resolved",
-                           help="free-form lifecycle status (default: resolved; "
-                                "a status starting with 'reopen' revives the item)")
-    p_resolve.add_argument("--note", help="optional context recorded on the event")
-    p_resolve.add_argument(
-        "--by", choices=["agent"],
-        help="declare an agent-initiated claim (requires --evidence); "
-             "omit for the human path (default)")
-    p_resolve.add_argument(
-        "--evidence",
-        help="verbatim transcript quote proving the claim — required with "
-             "--by agent, rejected without it")
-    p_resolve.add_argument("--project", help="project directory (default: DAIMON_PROJECT_DIR, then cwd)")
-    p_resolve.add_argument(
-        "--dry-run", action="store_true",
-        help="show what would resolve without writing an event — look before the write (#304)")
-    p_resolve.set_defaults(func=_cmd_resolve)
+    lifecycle.register(sub, fmt)
 
-    p_forget = sub.add_parser(
-        "forget", help="remove ONE item from the live checkpoint and record a "
-        "signed-state tombstone — content leaves disk and index; the event "
-        "carries only a hash (#321)",
-        epilog="Examples:\n  daimon forget o-3f8a2c --reason \"contains client name\"\n"
-               "  daimon forget \"wrong belief about retry nonce\" --dry-run\n",
-    )
-    p_forget.add_argument("target", help="item id (exact) or a query that must match exactly one item")
-    p_forget.add_argument("--reason", help="recorded on the tombstone event (redacted like any note)")
-    p_forget.add_argument("--project", help="project directory (default: DAIMON_PROJECT_DIR, then cwd)")
-    p_forget.add_argument(
-        "--dry-run", action="store_true",
-        help="show what would be forgotten without writing — look before a destructive op")
-    p_forget.set_defaults(func=_cmd_forget)
+    refute.register(sub, fmt)
 
-    p_reverify = sub.add_parser(
-        "reverify", help="evidence-gated reopen of a resolved item (#103) — "
-        "refuses without proof, so a claim can't get re-verified for free",
-        epilog="Examples:\n"
-               "  daimon reverify o-3f8a2c --evidence \"checked release page\"\n"
-               "  daimon reverify o-3f8a2c   # reopens only if the anchor still checks live\n"
-               "  daimon reverify o-3f8a2c   # rejects an unconfirmed supersede candidate — no evidence needed\n",
-    )
-    p_reverify.add_argument("target", help="item id (exact only — reverify is deliberate, no fuzzy match)")
-    p_reverify.add_argument("--evidence", help="why this claim can be trusted again")
-    p_reverify.add_argument("--project", help="project directory (default: DAIMON_PROJECT_DIR, then cwd)")
-    p_reverify.set_defaults(func=_cmd_reverify)
+    ruling.register(sub, fmt)
 
-    p_loops = sub.add_parser(
-        "loops", help="list open, briefable loop items with ids for this "
-        "project (#480) — the read counterpart to daimon resolve's write path",
-        epilog="Examples:\n  daimon loops\n  daimon loops --project .\n",
-    )
-    p_loops.add_argument("--project", help="project directory (default: DAIMON_PROJECT_DIR, then cwd)")
-    p_loops.set_defaults(func=_cmd_loops)
-
-    p_refute = sub.add_parser(
-        "refute",
-        help="author and query evidence-cited negative knowledge (#573)",
-        epilog="Examples:\n"
-               "  daimon refute list\n"
-               "  daimon refute search receipt verification\n"
-               "  daimon refute guard 'should we revisit #502?'\n",
-    )
-    refute_sub = p_refute.add_subparsers(dest="refute_cmd", required=True)
-    refute_sub.add_parser = functools.partial(
-        refute_sub.add_parser, formatter_class=fmt)
-
-    pr_add = refute_sub.add_parser(
-        "add", help="assert a scoped refutation; agent assertions stay candidates",
-        epilog="Example:\n"
-               "  daimon refute add --subject 'original #502 receipt design' "
-               "--verdict 'whole-file hashes do not prove span claims' "
-               "--scope 'carried-item receipt tiers' --anchor issue:502 "
-               "--evidence 'measurement:566/623 origin misses' "
-               "--ratify\n",
-    )
-    pr_add.add_argument("--subject", required=True, help="approach or claim rejected")
-    pr_add.add_argument("--verdict", required=True, help="what no longer holds")
-    pr_add.add_argument("--scope", required=True, help="where the verdict applies")
-    pr_add.add_argument(
-        "--evidence", action="append", required=True, metavar="SOURCE",
-        help="cited measurement, artifact, issue, or transcript source; recorded verbatim, not resolved or verified; repeatable")
-    pr_add.add_argument(
-        "--anchor", action="append", default=[], metavar="ANCHOR",
-        help="stable exact-match key such as issue:502 or command:daimon-why; repeatable")
-    pr_add.add_argument(
-        "--revisit-when", help="condition that makes reconsideration legitimate")
-    pr_add.add_argument("--by", choices=["agent"], default=None,
-                        help="declare yourself an agent; omit it only from an "
-                             "interactive terminal, which is the human path")
-    pr_add.add_argument(
-        "--ratify", action="store_true",
-        help="activate immediately; valid only on the human path, which is "
-             "an interactive terminal with no --by.")
-    pr_add.add_argument("--project", help="project directory (default: DAIMON_PROJECT_DIR, then cwd)")
-    pr_add.add_argument("--json", action="store_true", help="machine-readable output")
-    pr_add.set_defaults(func=_cmd_refute_add)
-
-    pr_ratify = refute_sub.add_parser(
-        "ratify", help="explicitly activate a candidate as a human decision")
-    pr_ratify.add_argument("refutation_id", help="exact r-… id")
-    pr_ratify.add_argument(
-        "--by", choices=["agent"], default=None,
-        help="declare yourself an agent; ratification then refuses, because "
-             "activation requires a human channel")
-    pr_ratify.add_argument("--note", help="optional ratification rationale")
-    pr_ratify.add_argument("--project", help="project directory (default: DAIMON_PROJECT_DIR, then cwd)")
-    pr_ratify.add_argument("--json", action="store_true", help="machine-readable output")
-    pr_ratify.set_defaults(func=_cmd_refute_ratify)
-
-    pr_revise = refute_sub.add_parser(
-        "revise", help="append a new evidence-cited version; inactive until ratified")
-    pr_revise.add_argument("refutation_id", help="exact r-… id")
-    pr_revise.add_argument("--subject", help="replacement subject")
-    pr_revise.add_argument("--verdict", help="replacement verdict")
-    pr_revise.add_argument("--scope", help="replacement scope")
-    pr_revise.add_argument(
-        "--anchor", action="append", default=None, metavar="ANCHOR",
-        help="replacement anchor set; repeatable")
-    pr_revise.add_argument("--revisit-when", help="replacement revisit condition")
-    pr_revise.add_argument(
-        "--evidence", action="append", required=True, metavar="SOURCE",
-        help="new evidence cited for the revision; recorded, not verified; repeatable")
-    pr_revise.add_argument("--by", choices=["agent"], default=None)
-    pr_revise.add_argument(
-        "--ratify", action="store_true",
-        help="activate the revision immediately; valid only on the human "
-             "path, which is an interactive terminal with no --by.")
-    pr_revise.add_argument("--project", help="project directory (default: DAIMON_PROJECT_DIR, then cwd)")
-    pr_revise.add_argument("--json", action="store_true", help="machine-readable output")
-    pr_revise.set_defaults(func=_cmd_refute_revise)
-
-    pr_overturn = refute_sub.add_parser(
-        "overturn",
-        help="cite evidence against the verdict; agent calls propose, humans deactivate")
-    pr_overturn.add_argument("refutation_id", help="exact r-… id")
-    pr_overturn.add_argument(
-        "--evidence", action="append", required=True, metavar="SOURCE",
-        help="evidence cited against the active verdict; recorded, not "
-             "verified; repeatable")
-    pr_overturn.add_argument("--note", help="optional explanation")
-    pr_overturn.add_argument("--by", choices=["agent"], default=None)
-    pr_overturn.add_argument("--project", help="project directory (default: DAIMON_PROJECT_DIR, then cwd)")
-    pr_overturn.add_argument("--json", action="store_true", help="machine-readable output")
-    pr_overturn.set_defaults(func=_cmd_refute_overturn)
-
-    pr_show = refute_sub.add_parser("show", help="show one refutation and its trust signals")
-    pr_show.add_argument("refutation_id", help="exact r-… id")
-    pr_show.add_argument("--project", help="project directory (default: DAIMON_PROJECT_DIR, then cwd)")
-    pr_show.add_argument("--json", action="store_true", help="machine-readable output")
-    pr_show.set_defaults(func=_cmd_refute_show)
-
-    pr_list = refute_sub.add_parser("list", help="list project refutations")
-    pr_list.add_argument("--state", action="append", choices=sorted(refutations.STATES),
-                         help="filter by state; repeatable")
-    pr_list.add_argument("--project", help="project directory (default: DAIMON_PROJECT_DIR, then cwd)")
-    pr_list.add_argument("--json", action="store_true", help="machine-readable output")
-    pr_list.set_defaults(func=_cmd_refute_list)
-
-    pr_search = refute_sub.add_parser(
-        "search", help="search the complete ledger without age decay")
-    pr_search.add_argument("query", nargs="+", help="subject, verdict, scope, or anchor terms")
-    pr_search.add_argument("--state", action="append", choices=sorted(refutations.STATES),
-                           help="filter by state; repeatable")
-    pr_search.add_argument("--project", help="project directory (default: DAIMON_PROJECT_DIR, then cwd)")
-    pr_search.add_argument("--json", action="store_true", help="machine-readable output")
-    pr_search.set_defaults(func=_cmd_refute_search)
-
-    pr_guard = refute_sub.add_parser(
-        "guard", help="check active refutations by exact anchor or subject phrase")
-    pr_guard.add_argument("query", nargs="+", help="the proposed approach or user prompt")
-    pr_guard.add_argument("--anchor", action="append", default=[], metavar="ANCHOR",
-                          help="candidate action anchor; repeatable")
-    pr_guard.add_argument("--project", help="project directory (default: DAIMON_PROJECT_DIR, then cwd)")
-    pr_guard.add_argument("--json", action="store_true", help="machine-readable output")
-    pr_guard.add_argument("--quiet", action="store_true",
-                          help="print nothing when no active refutation matches")
-    pr_guard.set_defaults(func=_cmd_refute_guard)
-
-    p_ruling = sub.add_parser(
-        "ruling",
-        help="standing rulings: human-ratified records that never decay (#693)",
-        epilog="Examples:\n"
-               "  daimon ruling propose --subject 'public posts' "
-               "--verdict 'internal numbers never appear in public posts' "
-               "--scope publishing --evidence issue:693 --by agent\n"
-               "  daimon ruling ratify r-1a2b3c4d5e6f\n",
-        formatter_class=argparse.RawDescriptionHelpFormatter)
-    ruling_sub = p_ruling.add_subparsers(dest="ruling_cmd", required=True)
-
-    rl_propose = ruling_sub.add_parser(
-        "propose", help="propose a standing ruling; agent proposals stay "
-                        "candidates until a human ratifies")
-    rl_propose.add_argument("--subject", required=True,
-                            help="what the ruling governs")
-    rl_propose.add_argument("--verdict", required=True,
-                            help="the ruling text itself (max 280 chars)")
-    rl_propose.add_argument("--scope", required=True,
-                            help="where it applies")
-    rl_propose.add_argument(
-        "--evidence", action="append", required=True, metavar="SOURCE",
-        help="cited source behind the ruling; recorded verbatim, not "
-             "resolved or verified; repeatable")
-    rl_propose.add_argument(
-        "--anchor", action="append", default=[], metavar="ANCHOR",
-        help="stable exact-match key such as issue:693; repeatable")
-    rl_propose.add_argument(
-        "--revisit-when", help="condition that makes reconsideration legitimate")
-    rl_propose.add_argument("--by", choices=["agent"], default=None,
-                            help="declare yourself an agent; omit it only "
-                                 "from an interactive terminal")
-    rl_propose.add_argument(
-        "--ratify", action="store_true",
-        help="activate immediately; valid only on the human path")
-    rl_propose.add_argument("--project", help="project directory (default: DAIMON_PROJECT_DIR, then cwd)")
-    rl_propose.add_argument("--json", action="store_true", help="machine-readable output")
-    rl_propose.set_defaults(func=_cmd_ruling_propose)
-
-    rl_ratify = ruling_sub.add_parser(
-        "ratify", help="activate a candidate ruling; prints the full text "
-                       "and confirms before the append")
-    rl_ratify.add_argument("ruling_id", help="exact r-… id")
-    rl_ratify.add_argument("--by", choices=["agent"], default=None,
-                           help="declare yourself an agent; ratification "
-                                "then refuses")
-    rl_ratify.add_argument("--note", help="optional ratification rationale")
-    rl_ratify.add_argument("--project", help="project directory (default: DAIMON_PROJECT_DIR, then cwd)")
-    rl_ratify.add_argument("--json", action="store_true", help="machine-readable output")
-    rl_ratify.set_defaults(func=_cmd_ruling_ratify)
-
-    rl_revise = ruling_sub.add_parser(
-        "revise", help="revise a ruling; on an active ruling an agent call "
-                       "records a proposal and the text stands")
-    rl_revise.add_argument("ruling_id", help="exact r-… id")
-    rl_revise.add_argument("--subject", help="replacement subject")
-    rl_revise.add_argument("--verdict", help="replacement ruling text")
-    rl_revise.add_argument("--scope", help="replacement scope")
-    rl_revise.add_argument(
-        "--anchor", action="append", default=None, metavar="ANCHOR",
-        help="replacement anchor set; repeatable")
-    rl_revise.add_argument("--revisit-when", help="replacement revisit condition")
-    rl_revise.add_argument(
-        "--evidence", action="append", required=True, metavar="SOURCE",
-        help="source cited for the revision; recorded, not verified; repeatable")
-    rl_revise.add_argument(
-        "--by", choices=["agent"], default=None,
-        help="declare yourself an agent; on an active ruling the call then "
-             "records a proposal and the text stands")
-    rl_revise.add_argument(
-        "--ratify", action="store_true",
-        help="refused: activation goes through `daimon ruling ratify`, "
-             "which shows the text before the write")
-    rl_revise.add_argument("--project", help="project directory (default: DAIMON_PROJECT_DIR, then cwd)")
-    rl_revise.add_argument("--json", action="store_true", help="machine-readable output")
-    rl_revise.set_defaults(func=_cmd_ruling_revise)
-
-    rl_retire = ruling_sub.add_parser(
-        "retire", help="end a ruling; agent calls propose, humans retire "
-                       "directly; evidence optional")
-    rl_retire.add_argument("ruling_id", help="exact r-… id")
-    rl_retire.add_argument(
-        "--evidence", action="append", default=None, metavar="SOURCE",
-        help="optional source for why the ruling stopped applying; repeatable")
-    rl_retire.add_argument("--note", help="optional explanation")
-    rl_retire.add_argument(
-        "--by", choices=["agent"], default=None,
-        help="declare yourself an agent; the call then records a retirement "
-             "proposal and the ruling stands until a human verdict")
-    rl_retire.add_argument("--project", help="project directory (default: DAIMON_PROJECT_DIR, then cwd)")
-    rl_retire.add_argument("--json", action="store_true", help="machine-readable output")
-    rl_retire.set_defaults(func=_cmd_ruling_retire)
-
-    rl_list = ruling_sub.add_parser("list", help="list project rulings")
-    rl_list.add_argument("--state", action="append",
-                         choices=sorted(refutations.STATES),
-                         help="filter by state; repeatable")
-    rl_list.add_argument("--project", help="project directory (default: DAIMON_PROJECT_DIR, then cwd)")
-    rl_list.add_argument("--json", action="store_true", help="machine-readable output")
-    rl_list.set_defaults(func=_cmd_ruling_list)
-
-    rl_show = ruling_sub.add_parser("show", help="show one ruling, including "
-                                                 "pending proposals")
-    rl_show.add_argument("ruling_id", help="exact r-… id")
-    rl_show.add_argument("--project", help="project directory (default: DAIMON_PROJECT_DIR, then cwd)")
-    rl_show.add_argument("--json", action="store_true", help="machine-readable output")
-    rl_show.set_defaults(func=_cmd_ruling_show)
-
-    p_amend = sub.add_parser(
-        "amend",
-        help="record an evidence-carrying state transition on a briefed item (#691)",
-        epilog="Examples:\n"
-               "  daimon amend o-1a2b3c4d5e6f --change progressed "
-               "--evidence 'the PR merged' --by agent\n"
-               "  daimon amend ratify a-0f1e2d3c4b5a\n",
-    )
-    amend_sub = p_amend.add_subparsers(dest="amend_cmd", required=True)
-    amend_sub.add_parser = functools.partial(
-        amend_sub.add_parser, formatter_class=fmt)
-
-    pa_prop = amend_sub.add_parser(
-        "propose",
-        help="propose an amendment; agent proposals stay candidates until "
-             "the session-end byte-check or a human verdict")
-    pa_prop.add_argument("item_id",
-                         help="exact item id from a briefing/loops handle")
-    pa_prop.add_argument(
-        "--change", required=True, choices=sorted(amendments.CHANGES),
-        help="the typed transition; the closed vocabulary is the render bound")
-    pa_prop.add_argument(
-        "--evidence", required=True,
-        help="verbatim transcript quote backing the change; byte-checked "
-             "against this session's transcript at session end")
-    pa_prop.add_argument("--note",
-                         help="short context; human channel only")
-    pa_prop.add_argument("--by", choices=["agent"], default=None,
-                         help="declare yourself an agent; omit it only from "
-                              "an interactive terminal, which is the human path")
-    pa_prop.add_argument("--project", help="project directory (default: DAIMON_PROJECT_DIR, then cwd)")
-    pa_prop.set_defaults(func=_cmd_amend_propose)
-
-    pa_ratify = amend_sub.add_parser(
-        "ratify", help="activate a candidate or verified amendment as a human decision")
-    pa_ratify.add_argument("amendment_id", help="exact a-… id")
-    pa_ratify.add_argument("--by", choices=["agent"], default=None,
-                           help="declare yourself an agent; ratification then "
-                                "refuses, because it requires a human channel")
-    pa_ratify.add_argument("--project", help="project directory (default: DAIMON_PROJECT_DIR, then cwd)")
-    pa_ratify.set_defaults(func=_cmd_amend_verdict)
-
-    pa_reject = amend_sub.add_parser(
-        "reject", help="reject an amendment with a reason, as a human decision")
-    pa_reject.add_argument("amendment_id", help="exact a-… id")
-    pa_reject.add_argument("--note", help="why it is wrong; kept on the record")
-    pa_reject.add_argument("--by", choices=["agent"], default=None,
-                           help="declare yourself an agent; rejection then "
-                                "refuses, because it requires a human channel")
-    pa_reject.add_argument("--project", help="project directory (default: DAIMON_PROJECT_DIR, then cwd)")
-    pa_reject.set_defaults(func=_cmd_amend_verdict)
-
-    pa_list = amend_sub.add_parser("list", help="list project amendments, candidates first")
-    pa_list.add_argument("--project", help="project directory (default: DAIMON_PROJECT_DIR, then cwd)")
-    pa_list.add_argument("--json", action="store_true", help="machine-readable output")
-    pa_list.set_defaults(func=_cmd_amend_list)
+    amend.register(sub, fmt)
 
     p_relations = sub.add_parser(
         "relations",
@@ -5146,53 +3037,7 @@ def build_parser() -> argparse.ArgumentParser:
         "--project", help="project directory (default: DAIMON_PROJECT_DIR, then cwd)")
     p_vr.set_defaults(func=_cmd_verify_receipt)
 
-    p_audit = sub.add_parser(
-        "audit",
-        help="read-only auditors that verify stored guarantees",
-    )
-    audit_sub = p_audit.add_subparsers(dest="audit_cmd", required=True)
-    pa_quotes = audit_sub.add_parser(
-        "quotes",
-        help="re-check stored verbatim quotes against their source transcripts "
-             "and report mismatches (read-only, never rewrites tags, #125)",
-        epilog="Examples:\n"
-               "  daimon audit quotes\n"
-               "  daimon audit quotes --all --top 20\n",
-    )
-    pa_quotes.add_argument(
-        "--project", help="project directory (default: DAIMON_PROJECT_DIR, then cwd)")
-    pa_quotes.add_argument(
-        "--all", action="store_true",
-        help="audit every project's checkpoints, not just the current one")
-    pa_quotes.add_argument(
-        "--top", type=int, default=10,
-        help="how many failing quotes to list (default: 10)")
-    pa_quotes.set_defaults(func=_cmd_audit_quotes)
-    pa_priv = audit_sub.add_parser(
-        "privacy",
-        help="prove no forgotten value's plaintext survives on any surface "
-             "(read-only; exit 0 clean, 1 residue, 3 cannot-prove)",
-        epilog="Examples:\n"
-               "  daimon audit privacy\n"
-               "  daimon audit privacy --all\n",
-    )
-    # Mutually exclusive: --project scopes to ONE bucket and --all audits every
-    # bucket, so together one of them is silently ignored — and the flag that
-    # loses decides which tombstone sets were checked. Fail loud instead.
-    pa_priv_scope = pa_priv.add_mutually_exclusive_group()
-    pa_priv_scope.add_argument(
-        "--project", help="project directory (default: DAIMON_PROJECT_DIR, then cwd)")
-    pa_priv_scope.add_argument(
-        "--all", action="store_true", dest="all_projects",
-        help="audit every local project, each against its own tombstone set")
-    pa_priv.set_defaults(func=_cmd_audit_privacy)
-    # Deprecated flat alias (#504-era name). metavar on the top-level
-    # subparsers (set below) keeps it out of the usage brace list.
-    p_audit_old = sub.add_parser("audit-quotes")
-    p_audit_old.add_argument("--project")
-    p_audit_old.add_argument("--all", action="store_true")
-    p_audit_old.add_argument("--top", type=int, default=10)
-    p_audit_old.set_defaults(func=_cmd_audit_quotes_deprecated)
+    audit.register(sub, fmt)
 
     p_heal = sub.add_parser(
         "heal",
@@ -5208,42 +3053,7 @@ def build_parser() -> argparse.ArgumentParser:
     )
     p_heal.set_defaults(func=_cmd_heal)
 
-    p_team = sub.add_parser(
-        "team", help="shared team memory: sidecar repo init/sync/status (#113)",
-        epilog="Examples:\n"
-               "  daimon team init git@github.com:org/team-memory.git\n"
-               "  daimon team sync\n"
-               "  daimon team status\n",
-    )
-    team_sub = p_team.add_subparsers(dest="team_cmd", required=True)
-    team_sub.add_parser = functools.partial(team_sub.add_parser, formatter_class=fmt)
-    pt_init = team_sub.add_parser(
-        "init", help="clone the private team sidecar repo (empty remote OK)"
-    )
-    pt_init.add_argument("remote_url", help="git remote URL of the PRIVATE team repo")
-    pt_init.set_defaults(func=_cmd_team_init)
-    pt_sync = team_sub.add_parser(
-        "sync", help="commit+push own checkpoints; fetch teammates' only on "
-                     "remote change (ls-remote gate)"
-    )
-    pt_sync.add_argument(
-        "--project",
-        help="accepted for CLI symmetry; sync is currently project-agnostic "
-             "(all own checkpoints sync regardless of project)",
-    )
-    pt_sync.add_argument(
-        "--apply-forget", action="store_true", dest="apply_forget",
-        help="also rewrite THIS machine's checkpoints under teammates' forget "
-             "tombstones (#600). Requires DAIMON_TEAM_APPLY_FORGET=1; typed "
-             "only — the SessionStart hook spawns a bare sync, so this can "
-             "never delete your belief state unattended",
-    )
-    pt_sync.set_defaults(func=_cmd_team_sync)
-    pt_status = team_sub.add_parser(
-        "status", help="per-remote freshness, own unpushed count, authors seen"
-    )
-    pt_status.add_argument("--json", action="store_true", help="machine-readable output")
-    pt_status.set_defaults(func=_cmd_team_status)
+    team.register(sub, fmt)
 
     p_cfg = sub.add_parser(
         "configure",
@@ -5295,52 +3105,9 @@ def build_parser() -> argparse.ArgumentParser:
     p_stats.add_argument("--json", action="store_true", help="machine-readable output")
     p_stats.set_defaults(func=_cmd_stats)
 
-    p_hooks = sub.add_parser(
-        "hooks",
-        help="ship host hook scripts from the package (#43): list, install, status",
-    )
-    hooks_sub = p_hooks.add_subparsers(dest="hooks_cmd", required=True)
-    hooks_sub.add_parser = functools.partial(hooks_sub.add_parser, formatter_class=fmt)
-    ph_list = hooks_sub.add_parser("list", help="hosts with packaged hook scripts")
-    ph_list.set_defaults(func=_cmd_hooks_list)
-    ph_status = hooks_sub.add_parser(
-        "status",
-        help="audit installed hook copies against the packaged versions "
-             "(CURRENT/STALE/MISSING/NOT INSTALLED); non-zero exit on drift",
-    )
-    ph_status.add_argument("--json", action="store_true", help="machine-readable output")
-    ph_status.set_defaults(func=_cmd_hooks_status)
-    ph_install = hooks_sub.add_parser(
-        "install",
-        help="copy a host's hook script(s) to the stable path ~/.daimon/hooks/ "
-             "and print the registration snippet — re-run after every upgrade",
-    )
-    ph_install.add_argument("host", help="host to install (see `daimon hooks list`)")
-    ph_install.set_defaults(func=_cmd_hooks_install)
+    hooks.register(sub, fmt)
 
-    p_skill = sub.add_parser(
-        "skill",
-        help="install the daimon agent skill into a host's rules/skills file (#66)")
-    skill_sub = p_skill.add_subparsers(dest="skill_cmd", required=True)
-    skill_sub.add_parser = functools.partial(skill_sub.add_parser, formatter_class=fmt)
-    ps_list = skill_sub.add_parser("list", help="hosts with a skill renderer")
-    ps_list.set_defaults(func=_cmd_skill_list)
-    ps_show = skill_sub.add_parser(
-        "show", help="print the canonical skill content")
-    ps_show.add_argument("--compact", action="store_true",
-                          help="print the rules-host variant instead of SKILL.md")
-    ps_show.set_defaults(func=_cmd_skill_show)
-    ps_install = skill_sub.add_parser(
-        "install", help="write the skill for a host (global scope by default)")
-    ps_install.add_argument("host", help="host to install (see `daimon skill list`)")
-    ps_install.add_argument("--project", action="store_true",
-                             help="write into the current repo instead of $HOME")
-    ps_install.set_defaults(func=_cmd_skill_install)
-    ps_uninstall = skill_sub.add_parser(
-        "uninstall", help="remove exactly what install wrote")
-    ps_uninstall.add_argument("host")
-    ps_uninstall.add_argument("--project", action="store_true")
-    ps_uninstall.set_defaults(func=_cmd_skill_uninstall)
+    skill.register(sub, fmt)
 
     p_mcp = sub.add_parser(
         "mcp",
